Crate brontes_classifier

Re-exports

pub use tree_builder::Classifier;

Modules

AaveV2

Generated by the following Solidity interface…

AaveV3

Generated by the following Solidity interface…

BalancerV1

Generated by the following Solidity interface…

CurveCryptoSwap

Generated by the following Solidity interface…

CurveCryptoSwapFactory

Generated by the following Solidity interface…

CurveTriCryptoFactory

Generated by the following Solidity interface…

CurveV1MetapoolFactory

Generated by the following Solidity interface…

CurveV2MetapoolFactory

Generated by the following Solidity interface…

CurvecrvUSDFactory

Generated by the following Solidity interface…

MakerPSM

Generated by the following Solidity interface…

PancakeSwapV3

Generated by the following Solidity interface…

SushiSwapV2

Generated by the following Solidity interface…

SushiSwapV3

Generated by the following Solidity interface…

UniswapV2

Generated by the following Solidity interface…

UniswapV2Factory

Generated by the following Solidity interface…

UniswapV3

Generated by the following Solidity interface…

UniswapV3Factory

Generated by the following Solidity interface…

UniswapX

Generated by the following Solidity interface…

aave

balancer

curve

maker

pancakeswap

sushiswap

transfer

tree_builder

uniswap

Structs

AaveV2flashLoanCall

AaveV2liquidationCallCall

AaveV3flashLoanCall

AaveV3flashLoanSimpleCall

AaveV3liquidationCallCall

BalancerV1swapExactAmountInCall

BalancerV1swapExactAmountOutCall

CurveCryptoSwapexchange_0Call

CurveCryptoSwapexchange_1Call

CurveCryptoSwapexchange_2Call

CurveCryptoSwapexchange_underlying_0Call

CurveV1MetapoolBaseDecoder

CurveV1MetapoolMetaDecoder

CurveV2MetapoolBaseDecoder

CurveV2MetapoolMetaDecoder0

CurveV2MetapoolMetaDecoder1

CurveV2MetapoolPlainDecoder0

CurveV2MetapoolPlainDecoder1

CurveV2MetapoolPlainDecoder2

DiscoveryProtocols

MakerPSMbuyGemCall

MakerPSMsellGemCall

PancakeSwapV3Decoder

PancakeSwapV3burnCall

PancakeSwapV3collectCall

PancakeSwapV3mintCall

PancakeSwapV3swapCall

ProtocolClassifications

SushiSwapV2Decoder

SushiSwapV2burnCall

SushiSwapV2mintCall

SushiSwapV2swapCall

SushiSwapV3Decoder

SushiSwapV3burnCall

SushiSwapV3collectCall

SushiSwapV3mintCall

SushiSwapV3swapCall

Transfer

Event with signature Transfer(address,address,uint256) and selector 0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef.

UniswapV2Decoder

UniswapV2burnCall

UniswapV2mintCall

UniswapV2swapCall

UniswapV3Decoder

UniswapV3burnCall

UniswapV3collectCall

UniswapV3mintCall

UniswapV3swapCall

UniswapXexecuteBatchCall

UniswapXexecuteBatchWithCallbackCall

UniswapXexecuteCall

UniswapXexecuteWithCallbackCall

coinsCall

Function with signature coins(uint256) and selector 0xc6610657.

coinsReturn

Container type for the return parameters of the coins(uint256) function.

decimalsCall

Function with signature decimals() and selector 0x313ce567.

decimalsReturn

Container type for the return parameters of the decimals() function.

nameCall

Function with signature name() and selector 0x06fdde03.

nameReturn

Container type for the return parameters of the name() function.

symbolCall

Function with signature symbol() and selector 0x95d89b41.

symbolReturn

Container type for the return parameters of the symbol() function.

totalSupplyCall

Function with signature totalSupply() and selector 0x18160ddd.

totalSupplyReturn

Container type for the return parameters of the totalSupply() function.

Constants

USDC_PSM_ADDRESS

USDP_PSM_ADDRESS

Traits

ActionCollection

FactoryDecoder

FactoryDecoderDispatch

IntoAction

Functions

__CurveV1MetapoolBaseDecoder_address_and_fn

__CurveV1MetapoolMetaDecoder_address_and_fn

__CurveV2MetapoolBaseDecoder_address_and_fn

__CurveV2MetapoolMetaDecoder0_address_and_fn

__CurveV2MetapoolMetaDecoder1_address_and_fn

__CurveV2MetapoolPlainDecoder0_address_and_fn

__CurveV2MetapoolPlainDecoder1_address_and_fn

__CurveV2MetapoolPlainDecoder2_address_and_fn

__PancakeSwapV3Decoder_address_and_fn

__SushiSwapV2Decoder_address_and_fn

__SushiSwapV3Decoder_address_and_fn

__UniswapV2Decoder_address_and_fn

__UniswapV3Decoder_address_and_fn

__action_sig_AaveV2flashLoanCall

__action_sig_AaveV2liquidationCallCall

__action_sig_AaveV3flashLoanCall

__action_sig_AaveV3flashLoanSimpleCall

__action_sig_AaveV3liquidationCallCall

__action_sig_BalancerV1swapExactAmountInCall

__action_sig_BalancerV1swapExactAmountOutCall

__action_sig_CurveCryptoSwapexchange_0Call

__action_sig_CurveCryptoSwapexchange_1Call

__action_sig_CurveCryptoSwapexchange_2Call

__action_sig_CurveCryptoSwapexchange_underlying_0Call

__action_sig_MakerPSMbuyGemCall

__action_sig_MakerPSMsellGemCall

__action_sig_PancakeSwapV3burnCall

__action_sig_PancakeSwapV3collectCall

__action_sig_PancakeSwapV3mintCall

__action_sig_PancakeSwapV3swapCall

__action_sig_SushiSwapV2burnCall

__action_sig_SushiSwapV2mintCall

__action_sig_SushiSwapV2swapCall

__action_sig_SushiSwapV3burnCall

__action_sig_SushiSwapV3collectCall

__action_sig_SushiSwapV3mintCall

__action_sig_SushiSwapV3swapCall

__action_sig_UniswapV2burnCall

__action_sig_UniswapV2mintCall

__action_sig_UniswapV2swapCall

__action_sig_UniswapV3burnCall

__action_sig_UniswapV3collectCall

__action_sig_UniswapV3mintCall

__action_sig_UniswapV3swapCall

__action_sig_UniswapXexecuteBatchCall

__action_sig_UniswapXexecuteBatchWithCallbackCall

__action_sig_UniswapXexecuteCall

__action_sig_UniswapXexecuteWithCallbackCall