Module brontes_classifier::AaveV3

Generated by the following Solidity interface…

interface AaveV3 {
    type InterestRateMode is uint8;
    struct EModeCategory {
        uint16 ltv;
        uint16 liquidationThreshold;
        uint16 liquidationBonus;
        address priceSource;
        string label;
    }
    struct ReserveConfigurationMap {
        uint256 data;
    }
    struct ReserveData {
        ReserveConfigurationMap configuration;
        uint128 liquidityIndex;
        uint128 currentLiquidityRate;
        uint128 variableBorrowIndex;
        uint128 currentVariableBorrowRate;
        uint128 currentStableBorrowRate;
        uint40 lastUpdateTimestamp;
        uint16 id;
        address aTokenAddress;
        address stableDebtTokenAddress;
        address variableDebtTokenAddress;
        address interestRateStrategyAddress;
        uint128 accruedToTreasury;
        uint128 unbacked;
        uint128 isolationModeTotalDebt;
    }
    struct UserConfigurationMap {
        uint256 data;
    }

    event BackUnbacked(address indexed reserve, address indexed backer, uint256 amount, uint256 fee);
    event Borrow(address indexed reserve, address user, address indexed onBehalfOf, uint256 amount, InterestRateMode interestRateMode, uint256 borrowRate, uint16 indexed referralCode);
    event FlashLoan(address indexed target, address initiator, address indexed asset, uint256 amount, InterestRateMode interestRateMode, uint256 premium, uint16 indexed referralCode);
    event IsolationModeTotalDebtUpdated(address indexed asset, uint256 totalDebt);
    event LiquidationCall(address indexed collateralAsset, address indexed debtAsset, address indexed user, uint256 debtToCover, uint256 liquidatedCollateralAmount, address liquidator, bool receiveAToken);
    event MintUnbacked(address indexed reserve, address user, address indexed onBehalfOf, uint256 amount, uint16 indexed referralCode);
    event MintedToTreasury(address indexed reserve, uint256 amountMinted);
    event RebalanceStableBorrowRate(address indexed reserve, address indexed user);
    event Repay(address indexed reserve, address indexed user, address indexed repayer, uint256 amount, bool useATokens);
    event ReserveDataUpdated(address indexed reserve, uint256 liquidityRate, uint256 stableBorrowRate, uint256 variableBorrowRate, uint256 liquidityIndex, uint256 variableBorrowIndex);
    event ReserveUsedAsCollateralDisabled(address indexed reserve, address indexed user);
    event ReserveUsedAsCollateralEnabled(address indexed reserve, address indexed user);
    event Supply(address indexed reserve, address user, address indexed onBehalfOf, uint256 amount, uint16 indexed referralCode);
    event SwapBorrowRateMode(address indexed reserve, address indexed user, InterestRateMode interestRateMode);
    event UserEModeSet(address indexed user, uint8 categoryId);
    event Withdraw(address indexed reserve, address indexed user, address indexed to, uint256 amount);

    constructor(address provider);

    function ADDRESSES_PROVIDER() external view returns (address);
    function BRIDGE_PROTOCOL_FEE() external view returns (uint256);
    function FLASHLOAN_PREMIUM_TOTAL() external view returns (uint128);
    function FLASHLOAN_PREMIUM_TO_PROTOCOL() external view returns (uint128);
    function MAX_NUMBER_RESERVES() external view returns (uint16);
    function MAX_STABLE_RATE_BORROW_SIZE_PERCENT() external view returns (uint256);
    function POOL_REVISION() external view returns (uint256);
    function backUnbacked(address asset, uint256 amount, uint256 fee) external;
    function borrow(address asset, uint256 amount, uint256 interestRateMode, uint16 referralCode, address onBehalfOf) external;
    function configureEModeCategory(uint8 id, EModeCategory memory category) external;
    function deposit(address asset, uint256 amount, address onBehalfOf, uint16 referralCode) external;
    function dropReserve(address asset) external;
    function finalizeTransfer(address asset, address from, address to, uint256 amount, uint256 balanceFromBefore, uint256 balanceToBefore) external;
    function flashLoan(address receiverAddress, address[] memory assets, uint256[] memory amounts, uint256[] memory interestRateModes, address onBehalfOf, bytes memory params, uint16 referralCode) external;
    function flashLoanSimple(address receiverAddress, address asset, uint256 amount, bytes memory params, uint16 referralCode) external;
    function getConfiguration(address asset) external view returns (ReserveConfigurationMap memory);
    function getEModeCategoryData(uint8 id) external view returns (EModeCategory memory);
    function getReserveAddressById(uint16 id) external view returns (address);
    function getReserveData(address asset) external view returns (ReserveData memory);
    function getReserveNormalizedIncome(address asset) external view returns (uint256);
    function getReserveNormalizedVariableDebt(address asset) external view returns (uint256);
    function getReservesList() external view returns (address[] memory);
    function getUserAccountData(address user) external view returns (uint256 totalCollateralBase, uint256 totalDebtBase, uint256 availableBorrowsBase, uint256 currentLiquidationThreshold, uint256 ltv, uint256 healthFactor);
    function getUserConfiguration(address user) external view returns (UserConfigurationMap memory);
    function getUserEMode(address user) external view returns (uint256);
    function initReserve(address asset, address aTokenAddress, address stableDebtAddress, address variableDebtAddress, address interestRateStrategyAddress) external;
    function initialize(address provider) external;
    function liquidationCall(address collateralAsset, address debtAsset, address user, uint256 debtToCover, bool receiveAToken) external;
    function mintToTreasury(address[] memory assets) external;
    function mintUnbacked(address asset, uint256 amount, address onBehalfOf, uint16 referralCode) external;
    function rebalanceStableBorrowRate(address asset, address user) external;
    function repay(address asset, uint256 amount, uint256 interestRateMode, address onBehalfOf) external returns (uint256);
    function repayWithATokens(address asset, uint256 amount, uint256 interestRateMode) external returns (uint256);
    function repayWithPermit(address asset, uint256 amount, uint256 interestRateMode, address onBehalfOf, uint256 deadline, uint8 permitV, bytes32 permitR, bytes32 permitS) external returns (uint256);
    function rescueTokens(address token, address to, uint256 amount) external;
    function resetIsolationModeTotalDebt(address asset) external;
    function setConfiguration(address asset, ReserveConfigurationMap memory configuration) external;
    function setReserveInterestRateStrategyAddress(address asset, address rateStrategyAddress) external;
    function setUserEMode(uint8 categoryId) external;
    function setUserUseReserveAsCollateral(address asset, bool useAsCollateral) external;
    function supply(address asset, uint256 amount, address onBehalfOf, uint16 referralCode) external;
    function supplyWithPermit(address asset, uint256 amount, address onBehalfOf, uint16 referralCode, uint256 deadline, uint8 permitV, bytes32 permitR, bytes32 permitS) external;
    function swapBorrowRateMode(address asset, uint256 interestRateMode) external;
    function updateBridgeProtocolFee(uint256 protocolFee) external;
    function updateFlashloanPremiums(uint128 flashLoanPremiumTotal, uint128 flashLoanPremiumToProtocol) external;
    function withdraw(address asset, uint256 amount, address to) external returns (uint256);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "provider",
        "type": "address",
        "internalType": "contract IPoolAddressesProvider"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "ADDRESSES_PROVIDER",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract IPoolAddressesProvider"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "BRIDGE_PROTOCOL_FEE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "FLASHLOAN_PREMIUM_TOTAL",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "FLASHLOAN_PREMIUM_TO_PROTOCOL",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MAX_NUMBER_RESERVES",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint16",
        "internalType": "uint16"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MAX_STABLE_RATE_BORROW_SIZE_PERCENT",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "POOL_REVISION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "backUnbacked",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "fee",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "borrow",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "interestRateMode",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "referralCode",
        "type": "uint16",
        "internalType": "uint16"
      },
      {
        "name": "onBehalfOf",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "configureEModeCategory",
    "inputs": [
      {
        "name": "id",
        "type": "uint8",
        "internalType": "uint8"
      },
      {
        "name": "category",
        "type": "tuple",
        "internalType": "struct DataTypes.EModeCategory",
        "components": [
          {
            "name": "ltv",
            "type": "uint16",
            "internalType": "uint16"
          },
          {
            "name": "liquidationThreshold",
            "type": "uint16",
            "internalType": "uint16"
          },
          {
            "name": "liquidationBonus",
            "type": "uint16",
            "internalType": "uint16"
          },
          {
            "name": "priceSource",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "label",
            "type": "string",
            "internalType": "string"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deposit",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "onBehalfOf",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "referralCode",
        "type": "uint16",
        "internalType": "uint16"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "dropReserve",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "finalizeTransfer",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "from",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "balanceFromBefore",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "balanceToBefore",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "flashLoan",
    "inputs": [
      {
        "name": "receiverAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "assets",
        "type": "address[]",
        "internalType": "address[]"
      },
      {
        "name": "amounts",
        "type": "uint256[]",
        "internalType": "uint256[]"
      },
      {
        "name": "interestRateModes",
        "type": "uint256[]",
        "internalType": "uint256[]"
      },
      {
        "name": "onBehalfOf",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "params",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "referralCode",
        "type": "uint16",
        "internalType": "uint16"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "flashLoanSimple",
    "inputs": [
      {
        "name": "receiverAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "params",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "referralCode",
        "type": "uint16",
        "internalType": "uint16"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "getConfiguration",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct DataTypes.ReserveConfigurationMap",
        "components": [
          {
            "name": "data",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getEModeCategoryData",
    "inputs": [
      {
        "name": "id",
        "type": "uint8",
        "internalType": "uint8"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct DataTypes.EModeCategory",
        "components": [
          {
            "name": "ltv",
            "type": "uint16",
            "internalType": "uint16"
          },
          {
            "name": "liquidationThreshold",
            "type": "uint16",
            "internalType": "uint16"
          },
          {
            "name": "liquidationBonus",
            "type": "uint16",
            "internalType": "uint16"
          },
          {
            "name": "priceSource",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "label",
            "type": "string",
            "internalType": "string"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getReserveAddressById",
    "inputs": [
      {
        "name": "id",
        "type": "uint16",
        "internalType": "uint16"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getReserveData",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct DataTypes.ReserveData",
        "components": [
          {
            "name": "configuration",
            "type": "tuple",
            "internalType": "struct DataTypes.ReserveConfigurationMap",
            "components": [
              {
                "name": "data",
                "type": "uint256",
                "internalType": "uint256"
              }
            ]
          },
          {
            "name": "liquidityIndex",
            "type": "uint128",
            "internalType": "uint128"
          },
          {
            "name": "currentLiquidityRate",
            "type": "uint128",
            "internalType": "uint128"
          },
          {
            "name": "variableBorrowIndex",
            "type": "uint128",
            "internalType": "uint128"
          },
          {
            "name": "currentVariableBorrowRate",
            "type": "uint128",
            "internalType": "uint128"
          },
          {
            "name": "currentStableBorrowRate",
            "type": "uint128",
            "internalType": "uint128"
          },
          {
            "name": "lastUpdateTimestamp",
            "type": "uint40",
            "internalType": "uint40"
          },
          {
            "name": "id",
            "type": "uint16",
            "internalType": "uint16"
          },
          {
            "name": "aTokenAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "stableDebtTokenAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "variableDebtTokenAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "interestRateStrategyAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "accruedToTreasury",
            "type": "uint128",
            "internalType": "uint128"
          },
          {
            "name": "unbacked",
            "type": "uint128",
            "internalType": "uint128"
          },
          {
            "name": "isolationModeTotalDebt",
            "type": "uint128",
            "internalType": "uint128"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getReserveNormalizedIncome",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getReserveNormalizedVariableDebt",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getReservesList",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address[]",
        "internalType": "address[]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getUserAccountData",
    "inputs": [
      {
        "name": "user",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "totalCollateralBase",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "totalDebtBase",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "availableBorrowsBase",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "currentLiquidationThreshold",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "ltv",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "healthFactor",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getUserConfiguration",
    "inputs": [
      {
        "name": "user",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct DataTypes.UserConfigurationMap",
        "components": [
          {
            "name": "data",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getUserEMode",
    "inputs": [
      {
        "name": "user",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "initReserve",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "aTokenAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "stableDebtAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "variableDebtAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "interestRateStrategyAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "initialize",
    "inputs": [
      {
        "name": "provider",
        "type": "address",
        "internalType": "contract IPoolAddressesProvider"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "liquidationCall",
    "inputs": [
      {
        "name": "collateralAsset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "debtAsset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "debtToCover",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "receiveAToken",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "mintToTreasury",
    "inputs": [
      {
        "name": "assets",
        "type": "address[]",
        "internalType": "address[]"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "mintUnbacked",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "onBehalfOf",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "referralCode",
        "type": "uint16",
        "internalType": "uint16"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "rebalanceStableBorrowRate",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "repay",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "interestRateMode",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "onBehalfOf",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "repayWithATokens",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "interestRateMode",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "repayWithPermit",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "interestRateMode",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "onBehalfOf",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "deadline",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "permitV",
        "type": "uint8",
        "internalType": "uint8"
      },
      {
        "name": "permitR",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "permitS",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "rescueTokens",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "resetIsolationModeTotalDebt",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setConfiguration",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "configuration",
        "type": "tuple",
        "internalType": "struct DataTypes.ReserveConfigurationMap",
        "components": [
          {
            "name": "data",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setReserveInterestRateStrategyAddress",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "rateStrategyAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setUserEMode",
    "inputs": [
      {
        "name": "categoryId",
        "type": "uint8",
        "internalType": "uint8"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setUserUseReserveAsCollateral",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "useAsCollateral",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "supply",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "onBehalfOf",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "referralCode",
        "type": "uint16",
        "internalType": "uint16"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "supplyWithPermit",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "onBehalfOf",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "referralCode",
        "type": "uint16",
        "internalType": "uint16"
      },
      {
        "name": "deadline",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "permitV",
        "type": "uint8",
        "internalType": "uint8"
      },
      {
        "name": "permitR",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "permitS",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "swapBorrowRateMode",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "interestRateMode",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "updateBridgeProtocolFee",
    "inputs": [
      {
        "name": "protocolFee",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "updateFlashloanPremiums",
    "inputs": [
      {
        "name": "flashLoanPremiumTotal",
        "type": "uint128",
        "internalType": "uint128"
      },
      {
        "name": "flashLoanPremiumToProtocol",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "withdraw",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "BackUnbacked",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "backer",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "fee",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Borrow",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "onBehalfOf",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "interestRateMode",
        "type": "uint8",
        "indexed": false,
        "internalType": "enum DataTypes.InterestRateMode"
      },
      {
        "name": "borrowRate",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "referralCode",
        "type": "uint16",
        "indexed": true,
        "internalType": "uint16"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "FlashLoan",
    "inputs": [
      {
        "name": "target",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "initiator",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "asset",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "interestRateMode",
        "type": "uint8",
        "indexed": false,
        "internalType": "enum DataTypes.InterestRateMode"
      },
      {
        "name": "premium",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "referralCode",
        "type": "uint16",
        "indexed": true,
        "internalType": "uint16"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "IsolationModeTotalDebtUpdated",
    "inputs": [
      {
        "name": "asset",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "totalDebt",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "LiquidationCall",
    "inputs": [
      {
        "name": "collateralAsset",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "debtAsset",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "debtToCover",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "liquidatedCollateralAmount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "liquidator",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "receiveAToken",
        "type": "bool",
        "indexed": false,
        "internalType": "bool"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "MintUnbacked",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "onBehalfOf",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "referralCode",
        "type": "uint16",
        "indexed": true,
        "internalType": "uint16"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "MintedToTreasury",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amountMinted",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RebalanceStableBorrowRate",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Repay",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "repayer",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "useATokens",
        "type": "bool",
        "indexed": false,
        "internalType": "bool"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ReserveDataUpdated",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "liquidityRate",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "stableBorrowRate",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "variableBorrowRate",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "liquidityIndex",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "variableBorrowIndex",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ReserveUsedAsCollateralDisabled",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ReserveUsedAsCollateralEnabled",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Supply",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "onBehalfOf",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "referralCode",
        "type": "uint16",
        "indexed": true,
        "internalType": "uint16"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "SwapBorrowRateMode",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "interestRateMode",
        "type": "uint8",
        "indexed": false,
        "internalType": "enum DataTypes.InterestRateMode"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "UserEModeSet",
    "inputs": [
      {
        "name": "user",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "categoryId",
        "type": "uint8",
        "indexed": false,
        "internalType": "uint8"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Withdraw",
    "inputs": [
      {
        "name": "reserve",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "user",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  }
]

Structs

ADDRESSES_PROVIDERCall

Function with signature ADDRESSES_PROVIDER() and selector 0x0542975c.

ADDRESSES_PROVIDERReturn

Container type for the return parameters of the ADDRESSES_PROVIDER() function.

BRIDGE_PROTOCOL_FEECall

Function with signature BRIDGE_PROTOCOL_FEE() and selector 0x272d9072.

BRIDGE_PROTOCOL_FEEReturn

Container type for the return parameters of the BRIDGE_PROTOCOL_FEE() function.

BackUnbacked

Event with signature BackUnbacked(address,address,uint256,uint256) and selector 0x281596e92b2d974beb7d4f124df30a0b39067b096893e95011ce4bdad798b759.

Borrow

Event with signature Borrow(address,address,address,uint256,uint8,uint256,uint16) and selector 0xb3d084820fb1a9decffb176436bd02558d15fac9b0ddfed8c465bc7359d7dce0.

EModeCategory

FLASHLOAN_PREMIUM_TOTALCall

Function with signature FLASHLOAN_PREMIUM_TOTAL() and selector 0x074b2e43.

FLASHLOAN_PREMIUM_TOTALReturn

Container type for the return parameters of the FLASHLOAN_PREMIUM_TOTAL() function.

FLASHLOAN_PREMIUM_TO_PROTOCOLCall

Function with signature FLASHLOAN_PREMIUM_TO_PROTOCOL() and selector 0x6a99c036.

FLASHLOAN_PREMIUM_TO_PROTOCOLReturn

Container type for the return parameters of the FLASHLOAN_PREMIUM_TO_PROTOCOL() function.

FlashLoan

Event with signature FlashLoan(address,address,address,uint256,uint8,uint256,uint16) and selector 0xefefaba5e921573100900a3ad9cf29f222d995fb3b6045797eaea7521bd8d6f0.

InterestRateMode

IsolationModeTotalDebtUpdated

Event with signature IsolationModeTotalDebtUpdated(address,uint256) and selector 0xaef84d3b40895fd58c561f3998000f0583abb992a52fbdc99ace8e8de4d676a5.

LiquidationCall

Event with signature LiquidationCall(address,address,address,uint256,uint256,address,bool) and selector 0xe413a321e8681d831f4dbccbca790d2952b56f977908e45be37335533e005286.

MAX_NUMBER_RESERVESCall

Function with signature MAX_NUMBER_RESERVES() and selector 0xf8119d51.

MAX_NUMBER_RESERVESReturn

Container type for the return parameters of the MAX_NUMBER_RESERVES() function.

MAX_STABLE_RATE_BORROW_SIZE_PERCENTCall

Function with signature MAX_STABLE_RATE_BORROW_SIZE_PERCENT() and selector 0xe82fec2f.

MAX_STABLE_RATE_BORROW_SIZE_PERCENTReturn

Container type for the return parameters of the MAX_STABLE_RATE_BORROW_SIZE_PERCENT() function.

MintUnbacked

Event with signature MintUnbacked(address,address,address,uint256,uint16) and selector 0xf25af37b3d3ec226063dc9bdc103ece7eb110a50f340fe854bb7bc1b0676d7d0.

MintedToTreasury

Event with signature MintedToTreasury(address,uint256) and selector 0xbfa21aa5d5f9a1f0120a95e7c0749f389863cbdbfff531aa7339077a5bc919de.

POOL_REVISIONCall

Function with signature POOL_REVISION() and selector 0x0148170e.

POOL_REVISIONReturn

Container type for the return parameters of the POOL_REVISION() function.

RebalanceStableBorrowRate

Event with signature RebalanceStableBorrowRate(address,address) and selector 0x9f439ae0c81e41a04d3fdfe07aed54e6a179fb0db15be7702eb66fa8ef6f5300.

Repay

Event with signature Repay(address,address,address,uint256,bool) and selector 0xa534c8dbe71f871f9f3530e97a74601fea17b426cae02e1c5aee42c96c784051.

ReserveConfigurationMap

ReserveData

ReserveDataUpdated

Event with signature ReserveDataUpdated(address,uint256,uint256,uint256,uint256,uint256) and selector 0x804c9b842b2748a22bb64b345453a3de7ca54a6ca45ce00d415894979e22897a.

ReserveUsedAsCollateralDisabled

Event with signature ReserveUsedAsCollateralDisabled(address,address) and selector 0x44c58d81365b66dd4b1a7f36c25aa97b8c71c361ee4937adc1a00000227db5dd.

ReserveUsedAsCollateralEnabled

Event with signature ReserveUsedAsCollateralEnabled(address,address) and selector 0x00058a56ea94653cdf4f152d227ace22d4c00ad99e2a43f58cb7d9e3feb295f2.

Supply

Event with signature Supply(address,address,address,uint256,uint16) and selector 0x2b627736bca15cd5381dcf80b0bf11fd197d01a037c52b927a881a10fb73ba61.

SwapBorrowRateMode

Event with signature SwapBorrowRateMode(address,address,uint8) and selector 0x7962b394d85a534033ba2efcf43cd36de57b7ebeb3de0ca4428965d9b3ddc481.

UserConfigurationMap

UserEModeSet

Event with signature UserEModeSet(address,uint8) and selector 0xd728da875fc88944cbf17638bcbe4af0eedaef63becd1d1c57cc097eb4608d84.

Withdraw

Event with signature Withdraw(address,address,address,uint256) and selector 0x3115d1449a7b732c986cba18244e897a450f61e1bb8d589cd2e69e6c8924f9f7.

backUnbackedCall

Function with signature backUnbacked(address,uint256,uint256) and selector 0xd65dc7a1.

backUnbackedReturn

Container type for the return parameters of the backUnbacked(address,uint256,uint256) function.

borrowCall

Function with signature borrow(address,uint256,uint256,uint16,address) and selector 0xa415bcad.

borrowReturn

Container type for the return parameters of the borrow(address,uint256,uint256,uint16,address) function.

configureEModeCategoryCall

Function with signature configureEModeCategory(uint8,(uint16,uint16,uint16,address,string)) and selector 0xd579ea7d.

configureEModeCategoryReturn

Container type for the return parameters of the configureEModeCategory(uint8,(uint16,uint16,uint16,address,string)) function.

constructorCall

Constructor`.

depositCall

Function with signature deposit(address,uint256,address,uint16) and selector 0xe8eda9df.

depositReturn

Container type for the return parameters of the deposit(address,uint256,address,uint16) function.

dropReserveCall

Function with signature dropReserve(address) and selector 0x63c9b860.

dropReserveReturn

Container type for the return parameters of the dropReserve(address) function.

finalizeTransferCall

Function with signature finalizeTransfer(address,address,address,uint256,uint256,uint256) and selector 0xd5ed3933.

finalizeTransferReturn

Container type for the return parameters of the finalizeTransfer(address,address,address,uint256,uint256,uint256) function.

flashLoanCall

Function with signature flashLoan(address,address[],uint256[],uint256[],address,bytes,uint16) and selector 0xab9c4b5d.

flashLoanReturn

Container type for the return parameters of the flashLoan(address,address[],uint256[],uint256[],address,bytes,uint16) function.

flashLoanSimpleCall

Function with signature flashLoanSimple(address,address,uint256,bytes,uint16) and selector 0x42b0b77c.

flashLoanSimpleReturn

Container type for the return parameters of the flashLoanSimple(address,address,uint256,bytes,uint16) function.

getConfigurationCall

Function with signature getConfiguration(address) and selector 0xc44b11f7.

getConfigurationReturn

Container type for the return parameters of the getConfiguration(address) function.

getEModeCategoryDataCall

Function with signature getEModeCategoryData(uint8) and selector 0x6c6f6ae1.

getEModeCategoryDataReturn

Container type for the return parameters of the getEModeCategoryData(uint8) function.

getReserveAddressByIdCall

Function with signature getReserveAddressById(uint16) and selector 0x52751797.

getReserveAddressByIdReturn

Container type for the return parameters of the getReserveAddressById(uint16) function.

getReserveDataCall

Function with signature getReserveData(address) and selector 0x35ea6a75.

getReserveDataReturn

Container type for the return parameters of the getReserveData(address) function.

getReserveNormalizedIncomeCall

Function with signature getReserveNormalizedIncome(address) and selector 0xd15e0053.

getReserveNormalizedIncomeReturn

Container type for the return parameters of the getReserveNormalizedIncome(address) function.

getReserveNormalizedVariableDebtCall

Function with signature getReserveNormalizedVariableDebt(address) and selector 0x386497fd.

getReserveNormalizedVariableDebtReturn

Container type for the return parameters of the getReserveNormalizedVariableDebt(address) function.

getReservesListCall

Function with signature getReservesList() and selector 0xd1946dbc.

getReservesListReturn

Container type for the return parameters of the getReservesList() function.

getUserAccountDataCall

Function with signature getUserAccountData(address) and selector 0xbf92857c.

getUserAccountDataReturn

Container type for the return parameters of the getUserAccountData(address) function.

getUserConfigurationCall

Function with signature getUserConfiguration(address) and selector 0x4417a583.

getUserConfigurationReturn

Container type for the return parameters of the getUserConfiguration(address) function.

getUserEModeCall

Function with signature getUserEMode(address) and selector 0xeddf1b79.

getUserEModeReturn

Container type for the return parameters of the getUserEMode(address) function.

initReserveCall

Function with signature initReserve(address,address,address,address,address) and selector 0x7a708e92.

initReserveReturn

Container type for the return parameters of the initReserve(address,address,address,address,address) function.

initializeCall

Function with signature initialize(address) and selector 0xc4d66de8.

initializeReturn

Container type for the return parameters of the initialize(address) function.

liquidationCallCall

Function with signature liquidationCall(address,address,address,uint256,bool) and selector 0x00a718a9.

liquidationCallReturn

Container type for the return parameters of the liquidationCall(address,address,address,uint256,bool) function.

mintToTreasuryCall

Function with signature mintToTreasury(address[]) and selector 0x9cd19996.

mintToTreasuryReturn

Container type for the return parameters of the mintToTreasury(address[]) function.

mintUnbackedCall

Function with signature mintUnbacked(address,uint256,address,uint16) and selector 0x69a933a5.

mintUnbackedReturn

Container type for the return parameters of the mintUnbacked(address,uint256,address,uint16) function.

rebalanceStableBorrowRateCall

Function with signature rebalanceStableBorrowRate(address,address) and selector 0xcd112382.

rebalanceStableBorrowRateReturn

Container type for the return parameters of the rebalanceStableBorrowRate(address,address) function.

repayCall

Function with signature repay(address,uint256,uint256,address) and selector 0x573ade81.

repayReturn

Container type for the return parameters of the repay(address,uint256,uint256,address) function.

repayWithATokensCall

Function with signature repayWithATokens(address,uint256,uint256) and selector 0x2dad97d4.

repayWithATokensReturn

Container type for the return parameters of the repayWithATokens(address,uint256,uint256) function.

repayWithPermitCall

Function with signature repayWithPermit(address,uint256,uint256,address,uint256,uint8,bytes32,bytes32) and selector 0xee3e210b.

repayWithPermitReturn

Container type for the return parameters of the repayWithPermit(address,uint256,uint256,address,uint256,uint8,bytes32,bytes32) function.

rescueTokensCall

Function with signature rescueTokens(address,address,uint256) and selector 0xcea9d26f.

rescueTokensReturn

Container type for the return parameters of the rescueTokens(address,address,uint256) function.

resetIsolationModeTotalDebtCall

Function with signature resetIsolationModeTotalDebt(address) and selector 0xe43e88a1.

resetIsolationModeTotalDebtReturn

Container type for the return parameters of the resetIsolationModeTotalDebt(address) function.

setConfigurationCall

Function with signature setConfiguration(address,(uint256)) and selector 0xf51e435b.

setConfigurationReturn

Container type for the return parameters of the setConfiguration(address,(uint256)) function.

setReserveInterestRateStrategyAddressCall

Function with signature setReserveInterestRateStrategyAddress(address,address) and selector 0x1d2118f9.

setReserveInterestRateStrategyAddressReturn

Container type for the return parameters of the setReserveInterestRateStrategyAddress(address,address) function.

setUserEModeCall

Function with signature setUserEMode(uint8) and selector 0x28530a47.

setUserEModeReturn

Container type for the return parameters of the setUserEMode(uint8) function.

setUserUseReserveAsCollateralCall

Function with signature setUserUseReserveAsCollateral(address,bool) and selector 0x5a3b74b9.

setUserUseReserveAsCollateralReturn

Container type for the return parameters of the setUserUseReserveAsCollateral(address,bool) function.

supplyCall

Function with signature supply(address,uint256,address,uint16) and selector 0x617ba037.

supplyReturn

Container type for the return parameters of the supply(address,uint256,address,uint16) function.

supplyWithPermitCall

Function with signature supplyWithPermit(address,uint256,address,uint16,uint256,uint8,bytes32,bytes32) and selector 0x02c205f0.

supplyWithPermitReturn

Container type for the return parameters of the supplyWithPermit(address,uint256,address,uint16,uint256,uint8,bytes32,bytes32) function.

swapBorrowRateModeCall

Function with signature swapBorrowRateMode(address,uint256) and selector 0x94ba89a2.

swapBorrowRateModeReturn

Container type for the return parameters of the swapBorrowRateMode(address,uint256) function.

updateBridgeProtocolFeeCall

Function with signature updateBridgeProtocolFee(uint256) and selector 0x3036b439.

updateBridgeProtocolFeeReturn

Container type for the return parameters of the updateBridgeProtocolFee(uint256) function.

updateFlashloanPremiumsCall

Function with signature updateFlashloanPremiums(uint128,uint128) and selector 0xbcb6e522.

updateFlashloanPremiumsReturn

Container type for the return parameters of the updateFlashloanPremiums(uint128,uint128) function.

withdrawCall

Function with signature withdraw(address,uint256,address) and selector 0x69328dec.

withdrawReturn

Container type for the return parameters of the withdraw(address,uint256,address) function.

Enums

AaveV3Calls

Container for all the AaveV3 function calls.

AaveV3Events

Container for all the AaveV3 events.