Module brontes_classifier::BalancerV1

Generated by the following Solidity interface…

interface BalancerV1 {
    event Approval(address indexed src, address indexed dst, uint256 amt);
    event LOG_CALL(bytes4 indexed sig, address indexed caller, bytes data) anonymous;
    event LOG_EXIT(address indexed caller, address indexed tokenOut, uint256 tokenAmountOut);
    event LOG_JOIN(address indexed caller, address indexed tokenIn, uint256 tokenAmountIn);
    event LOG_SWAP(address indexed caller, address indexed tokenIn, address indexed tokenOut, uint256 tokenAmountIn, uint256 tokenAmountOut);
    event Transfer(address indexed src, address indexed dst, uint256 amt);

    constructor();

    function BONE() external view returns (uint256);
    function BPOW_PRECISION() external view returns (uint256);
    function EXIT_FEE() external view returns (uint256);
    function INIT_POOL_SUPPLY() external view returns (uint256);
    function MAX_BOUND_TOKENS() external view returns (uint256);
    function MAX_BPOW_BASE() external view returns (uint256);
    function MAX_FEE() external view returns (uint256);
    function MAX_IN_RATIO() external view returns (uint256);
    function MAX_OUT_RATIO() external view returns (uint256);
    function MAX_TOTAL_WEIGHT() external view returns (uint256);
    function MAX_WEIGHT() external view returns (uint256);
    function MIN_BALANCE() external view returns (uint256);
    function MIN_BOUND_TOKENS() external view returns (uint256);
    function MIN_BPOW_BASE() external view returns (uint256);
    function MIN_FEE() external view returns (uint256);
    function MIN_WEIGHT() external view returns (uint256);
    function allowance(address src, address dst) external view returns (uint256);
    function approve(address dst, uint256 amt) external returns (bool);
    function balanceOf(address whom) external view returns (uint256);
    function bind(address token, uint256 balance, uint256 denorm) external;
    function calcInGivenOut(uint256 tokenBalanceIn, uint256 tokenWeightIn, uint256 tokenBalanceOut, uint256 tokenWeightOut, uint256 tokenAmountOut, uint256 swapFee) external view returns (uint256 tokenAmountIn);
    function calcOutGivenIn(uint256 tokenBalanceIn, uint256 tokenWeightIn, uint256 tokenBalanceOut, uint256 tokenWeightOut, uint256 tokenAmountIn, uint256 swapFee) external view returns (uint256 tokenAmountOut);
    function calcPoolInGivenSingleOut(uint256 tokenBalanceOut, uint256 tokenWeightOut, uint256 poolSupply, uint256 totalWeight, uint256 tokenAmountOut, uint256 swapFee) external view returns (uint256 poolAmountIn);
    function calcPoolOutGivenSingleIn(uint256 tokenBalanceIn, uint256 tokenWeightIn, uint256 poolSupply, uint256 totalWeight, uint256 tokenAmountIn, uint256 swapFee) external view returns (uint256 poolAmountOut);
    function calcSingleInGivenPoolOut(uint256 tokenBalanceIn, uint256 tokenWeightIn, uint256 poolSupply, uint256 totalWeight, uint256 poolAmountOut, uint256 swapFee) external view returns (uint256 tokenAmountIn);
    function calcSingleOutGivenPoolIn(uint256 tokenBalanceOut, uint256 tokenWeightOut, uint256 poolSupply, uint256 totalWeight, uint256 poolAmountIn, uint256 swapFee) external view returns (uint256 tokenAmountOut);
    function calcSpotPrice(uint256 tokenBalanceIn, uint256 tokenWeightIn, uint256 tokenBalanceOut, uint256 tokenWeightOut, uint256 swapFee) external view returns (uint256 spotPrice);
    function decimals() external view returns (uint8);
    function decreaseApproval(address dst, uint256 amt) external returns (bool);
    function exitPool(uint256 poolAmountIn, uint256[] memory minAmountsOut) external;
    function exitswapExternAmountOut(address tokenOut, uint256 tokenAmountOut, uint256 maxPoolAmountIn) external returns (uint256 poolAmountIn);
    function exitswapPoolAmountIn(address tokenOut, uint256 poolAmountIn, uint256 minAmountOut) external returns (uint256 tokenAmountOut);
    function finalize() external;
    function getBalance(address token) external view returns (uint256);
    function getColor() external view returns (bytes32);
    function getController() external view returns (address);
    function getCurrentTokens() external view returns (address[] memory tokens);
    function getDenormalizedWeight(address token) external view returns (uint256);
    function getFinalTokens() external view returns (address[] memory tokens);
    function getNormalizedWeight(address token) external view returns (uint256);
    function getNumTokens() external view returns (uint256);
    function getSpotPrice(address tokenIn, address tokenOut) external view returns (uint256 spotPrice);
    function getSpotPriceSansFee(address tokenIn, address tokenOut) external view returns (uint256 spotPrice);
    function getSwapFee() external view returns (uint256);
    function getTotalDenormalizedWeight() external view returns (uint256);
    function gulp(address token) external;
    function increaseApproval(address dst, uint256 amt) external returns (bool);
    function isBound(address t) external view returns (bool);
    function isFinalized() external view returns (bool);
    function isPublicSwap() external view returns (bool);
    function joinPool(uint256 poolAmountOut, uint256[] memory maxAmountsIn) external;
    function joinswapExternAmountIn(address tokenIn, uint256 tokenAmountIn, uint256 minPoolAmountOut) external returns (uint256 poolAmountOut);
    function joinswapPoolAmountOut(address tokenIn, uint256 poolAmountOut, uint256 maxAmountIn) external returns (uint256 tokenAmountIn);
    function name() external view returns (string memory);
    function rebind(address token, uint256 balance, uint256 denorm) external;
    function setController(address manager) external;
    function setPublicSwap(bool public_) external;
    function setSwapFee(uint256 swapFee) external;
    function swapExactAmountIn(address tokenIn, uint256 tokenAmountIn, address tokenOut, uint256 minAmountOut, uint256 maxPrice) external returns (uint256 tokenAmountOut, uint256 spotPriceAfter);
    function swapExactAmountOut(address tokenIn, uint256 maxAmountIn, address tokenOut, uint256 tokenAmountOut, uint256 maxPrice) external returns (uint256 tokenAmountIn, uint256 spotPriceAfter);
    function symbol() external view returns (string memory);
    function totalSupply() external view returns (uint256);
    function transfer(address dst, uint256 amt) external returns (bool);
    function transferFrom(address src, address dst, uint256 amt) external returns (bool);
    function unbind(address token) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "BONE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "BPOW_PRECISION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "EXIT_FEE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "INIT_POOL_SUPPLY",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MAX_BOUND_TOKENS",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MAX_BPOW_BASE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MAX_FEE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MAX_IN_RATIO",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MAX_OUT_RATIO",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MAX_TOTAL_WEIGHT",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MAX_WEIGHT",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MIN_BALANCE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MIN_BOUND_TOKENS",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MIN_BPOW_BASE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MIN_FEE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MIN_WEIGHT",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "allowance",
    "inputs": [
      {
        "name": "src",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "dst",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "approve",
    "inputs": [
      {
        "name": "dst",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amt",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "balanceOf",
    "inputs": [
      {
        "name": "whom",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "bind",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "balance",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "denorm",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "calcInGivenOut",
    "inputs": [
      {
        "name": "tokenBalanceIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenWeightIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenBalanceOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenWeightOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "swapFee",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "tokenAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "calcOutGivenIn",
    "inputs": [
      {
        "name": "tokenBalanceIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenWeightIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenBalanceOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenWeightOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "swapFee",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "tokenAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "calcPoolInGivenSingleOut",
    "inputs": [
      {
        "name": "tokenBalanceOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenWeightOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "poolSupply",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "totalWeight",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "swapFee",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "poolAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "calcPoolOutGivenSingleIn",
    "inputs": [
      {
        "name": "tokenBalanceIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenWeightIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "poolSupply",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "totalWeight",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "swapFee",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "poolAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "calcSingleInGivenPoolOut",
    "inputs": [
      {
        "name": "tokenBalanceIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenWeightIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "poolSupply",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "totalWeight",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "poolAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "swapFee",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "tokenAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "calcSingleOutGivenPoolIn",
    "inputs": [
      {
        "name": "tokenBalanceOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenWeightOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "poolSupply",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "totalWeight",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "poolAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "swapFee",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "tokenAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "calcSpotPrice",
    "inputs": [
      {
        "name": "tokenBalanceIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenWeightIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenBalanceOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenWeightOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "swapFee",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "spotPrice",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "decimals",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint8",
        "internalType": "uint8"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "decreaseApproval",
    "inputs": [
      {
        "name": "dst",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amt",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "exitPool",
    "inputs": [
      {
        "name": "poolAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "minAmountsOut",
        "type": "uint256[]",
        "internalType": "uint256[]"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "exitswapExternAmountOut",
    "inputs": [
      {
        "name": "tokenOut",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "maxPoolAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "poolAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "exitswapPoolAmountIn",
    "inputs": [
      {
        "name": "tokenOut",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "poolAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "minAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "tokenAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "finalize",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "getBalance",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getColor",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getController",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getCurrentTokens",
    "inputs": [],
    "outputs": [
      {
        "name": "tokens",
        "type": "address[]",
        "internalType": "address[]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getDenormalizedWeight",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getFinalTokens",
    "inputs": [],
    "outputs": [
      {
        "name": "tokens",
        "type": "address[]",
        "internalType": "address[]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getNormalizedWeight",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getNumTokens",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getSpotPrice",
    "inputs": [
      {
        "name": "tokenIn",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenOut",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "spotPrice",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getSpotPriceSansFee",
    "inputs": [
      {
        "name": "tokenIn",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenOut",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "spotPrice",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getSwapFee",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getTotalDenormalizedWeight",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "gulp",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "increaseApproval",
    "inputs": [
      {
        "name": "dst",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amt",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "isBound",
    "inputs": [
      {
        "name": "t",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "isFinalized",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "isPublicSwap",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "joinPool",
    "inputs": [
      {
        "name": "poolAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "maxAmountsIn",
        "type": "uint256[]",
        "internalType": "uint256[]"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "joinswapExternAmountIn",
    "inputs": [
      {
        "name": "tokenIn",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "minPoolAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "poolAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "joinswapPoolAmountOut",
    "inputs": [
      {
        "name": "tokenIn",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "poolAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "maxAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "tokenAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "name",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "rebind",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "balance",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "denorm",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setController",
    "inputs": [
      {
        "name": "manager",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setPublicSwap",
    "inputs": [
      {
        "name": "public_",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setSwapFee",
    "inputs": [
      {
        "name": "swapFee",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "swapExactAmountIn",
    "inputs": [
      {
        "name": "tokenIn",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenOut",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "minAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "maxPrice",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "tokenAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "spotPriceAfter",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "swapExactAmountOut",
    "inputs": [
      {
        "name": "tokenIn",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "maxAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokenOut",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenAmountOut",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "maxPrice",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "tokenAmountIn",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "spotPriceAfter",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "symbol",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "totalSupply",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "transfer",
    "inputs": [
      {
        "name": "dst",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amt",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferFrom",
    "inputs": [
      {
        "name": "src",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "dst",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amt",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "unbind",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "Approval",
    "inputs": [
      {
        "name": "src",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "dst",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amt",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "LOG_CALL",
    "inputs": [
      {
        "name": "sig",
        "type": "bytes4",
        "indexed": true,
        "internalType": "bytes4"
      },
      {
        "name": "caller",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "data",
        "type": "bytes",
        "indexed": false,
        "internalType": "bytes"
      }
    ],
    "anonymous": true
  },
  {
    "type": "event",
    "name": "LOG_EXIT",
    "inputs": [
      {
        "name": "caller",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "tokenOut",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "tokenAmountOut",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "LOG_JOIN",
    "inputs": [
      {
        "name": "caller",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "tokenIn",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "tokenAmountIn",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "LOG_SWAP",
    "inputs": [
      {
        "name": "caller",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "tokenIn",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "tokenOut",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "tokenAmountIn",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "tokenAmountOut",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Transfer",
    "inputs": [
      {
        "name": "src",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "dst",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amt",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  }
]

Structs

Approval

Event with signature Approval(address,address,uint256) and selector 0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925.

BONECall

Function with signature BONE() and selector 0xc36596a6.

BONEReturn

Container type for the return parameters of the BONE() function.

BPOW_PRECISIONCall

Function with signature BPOW_PRECISION() and selector 0x189d00ca.

BPOW_PRECISIONReturn

Container type for the return parameters of the BPOW_PRECISION() function.

EXIT_FEECall

Function with signature EXIT_FEE() and selector 0xc6580d12.

EXIT_FEEReturn

Container type for the return parameters of the EXIT_FEE() function.

INIT_POOL_SUPPLYCall

Function with signature INIT_POOL_SUPPLY() and selector 0x9381cd2b.

INIT_POOL_SUPPLYReturn

Container type for the return parameters of the INIT_POOL_SUPPLY() function.

LOG_CALL

Event with signature LOG_CALL(bytes4,address,bytes) and selector 0x25fce1fe01d9b241fda40b2152ddd6f4ba063fcfb3c2c81dddf84ee20d3f341f.

LOG_EXIT

Event with signature LOG_EXIT(address,address,uint256) and selector 0xe74c91552b64c2e2e7bd255639e004e693bd3e1d01cc33e65610b86afcc1ffed.

LOG_JOIN

Event with signature LOG_JOIN(address,address,uint256) and selector 0x63982df10efd8dfaaaa0fcc7f50b2d93b7cba26ccc48adee2873220d485dc39a.

LOG_SWAP

Event with signature LOG_SWAP(address,address,address,uint256,uint256) and selector 0x908fb5ee8f16c6bc9bc3690973819f32a4d4b10188134543c88706e0e1d43378.

MAX_BOUND_TOKENSCall

Function with signature MAX_BOUND_TOKENS() and selector 0xb0e0d136.

MAX_BOUND_TOKENSReturn

Container type for the return parameters of the MAX_BOUND_TOKENS() function.

MAX_BPOW_BASECall

Function with signature MAX_BPOW_BASE() and selector 0xbc694ea2.

MAX_BPOW_BASEReturn

Container type for the return parameters of the MAX_BPOW_BASE() function.

MAX_FEECall

Function with signature MAX_FEE() and selector 0xbc063e1a.

MAX_FEEReturn

Container type for the return parameters of the MAX_FEE() function.

MAX_IN_RATIOCall

Function with signature MAX_IN_RATIO() and selector 0xec093021.

MAX_IN_RATIOReturn

Container type for the return parameters of the MAX_IN_RATIO() function.

MAX_OUT_RATIOCall

Function with signature MAX_OUT_RATIO() and selector 0x992e2a92.

MAX_OUT_RATIOReturn

Container type for the return parameters of the MAX_OUT_RATIO() function.

MAX_TOTAL_WEIGHTCall

Function with signature MAX_TOTAL_WEIGHT() and selector 0x09a3bbe4.

MAX_TOTAL_WEIGHTReturn

Container type for the return parameters of the MAX_TOTAL_WEIGHT() function.

MAX_WEIGHTCall

Function with signature MAX_WEIGHT() and selector 0xe4a28a52.

MAX_WEIGHTReturn

Container type for the return parameters of the MAX_WEIGHT() function.

MIN_BALANCECall

Function with signature MIN_BALANCE() and selector 0x867378c5.

MIN_BALANCEReturn

Container type for the return parameters of the MIN_BALANCE() function.

MIN_BOUND_TOKENSCall

Function with signature MIN_BOUND_TOKENS() and selector 0xb7b800a4.

MIN_BOUND_TOKENSReturn

Container type for the return parameters of the MIN_BOUND_TOKENS() function.

MIN_BPOW_BASECall

Function with signature MIN_BPOW_BASE() and selector 0xba019dab.

MIN_BPOW_BASEReturn

Container type for the return parameters of the MIN_BPOW_BASE() function.

MIN_FEECall

Function with signature MIN_FEE() and selector 0x76c7a3c7.

MIN_FEEReturn

Container type for the return parameters of the MIN_FEE() function.

MIN_WEIGHTCall

Function with signature MIN_WEIGHT() and selector 0x218b5382.

MIN_WEIGHTReturn

Container type for the return parameters of the MIN_WEIGHT() function.

Transfer

Event with signature Transfer(address,address,uint256) and selector 0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef.

allowanceCall

Function with signature allowance(address,address) and selector 0xdd62ed3e.

allowanceReturn

Container type for the return parameters of the allowance(address,address) function.

approveCall

Function with signature approve(address,uint256) and selector 0x095ea7b3.

approveReturn

Container type for the return parameters of the approve(address,uint256) function.

balanceOfCall

Function with signature balanceOf(address) and selector 0x70a08231.

balanceOfReturn

Container type for the return parameters of the balanceOf(address) function.

bindCall

Function with signature bind(address,uint256,uint256) and selector 0xe4e1e538.

bindReturn

Container type for the return parameters of the bind(address,uint256,uint256) function.

calcInGivenOutCall

Function with signature calcInGivenOut(uint256,uint256,uint256,uint256,uint256,uint256) and selector 0xf8d6aed4.

calcInGivenOutReturn

Container type for the return parameters of the calcInGivenOut(uint256,uint256,uint256,uint256,uint256,uint256) function.

calcOutGivenInCall

Function with signature calcOutGivenIn(uint256,uint256,uint256,uint256,uint256,uint256) and selector 0xba9530a6.

calcOutGivenInReturn

Container type for the return parameters of the calcOutGivenIn(uint256,uint256,uint256,uint256,uint256,uint256) function.

calcPoolInGivenSingleOutCall

Function with signature calcPoolInGivenSingleOut(uint256,uint256,uint256,uint256,uint256,uint256) and selector 0x82f652ad.

calcPoolInGivenSingleOutReturn

Container type for the return parameters of the calcPoolInGivenSingleOut(uint256,uint256,uint256,uint256,uint256,uint256) function.

calcPoolOutGivenSingleInCall

Function with signature calcPoolOutGivenSingleIn(uint256,uint256,uint256,uint256,uint256,uint256) and selector 0x8656b653.

calcPoolOutGivenSingleInReturn

Container type for the return parameters of the calcPoolOutGivenSingleIn(uint256,uint256,uint256,uint256,uint256,uint256) function.

calcSingleInGivenPoolOutCall

Function with signature calcSingleInGivenPoolOut(uint256,uint256,uint256,uint256,uint256,uint256) and selector 0x5c1bbaf7.

calcSingleInGivenPoolOutReturn

Container type for the return parameters of the calcSingleInGivenPoolOut(uint256,uint256,uint256,uint256,uint256,uint256) function.

calcSingleOutGivenPoolInCall

Function with signature calcSingleOutGivenPoolIn(uint256,uint256,uint256,uint256,uint256,uint256) and selector 0x89298012.

calcSingleOutGivenPoolInReturn

Container type for the return parameters of the calcSingleOutGivenPoolIn(uint256,uint256,uint256,uint256,uint256,uint256) function.

calcSpotPriceCall

Function with signature calcSpotPrice(uint256,uint256,uint256,uint256,uint256) and selector 0xa221ee49.

calcSpotPriceReturn

Container type for the return parameters of the calcSpotPrice(uint256,uint256,uint256,uint256,uint256) function.

constructorCall

Constructor`.

decimalsCall

Function with signature decimals() and selector 0x313ce567.

decimalsReturn

Container type for the return parameters of the decimals() function.

decreaseApprovalCall

Function with signature decreaseApproval(address,uint256) and selector 0x66188463.

decreaseApprovalReturn

Container type for the return parameters of the decreaseApproval(address,uint256) function.

exitPoolCall

Function with signature exitPool(uint256,uint256[]) and selector 0xb02f0b73.

exitPoolReturn

Container type for the return parameters of the exitPool(uint256,uint256[]) function.

exitswapExternAmountOutCall

Function with signature exitswapExternAmountOut(address,uint256,uint256) and selector 0x02c96748.

exitswapExternAmountOutReturn

Container type for the return parameters of the exitswapExternAmountOut(address,uint256,uint256) function.

exitswapPoolAmountInCall

Function with signature exitswapPoolAmountIn(address,uint256,uint256) and selector 0x46ab38f1.

exitswapPoolAmountInReturn

Container type for the return parameters of the exitswapPoolAmountIn(address,uint256,uint256) function.

finalizeCall

Function with signature finalize() and selector 0x4bb278f3.

finalizeReturn

Container type for the return parameters of the finalize() function.

getBalanceCall

Function with signature getBalance(address) and selector 0xf8b2cb4f.

getBalanceReturn

Container type for the return parameters of the getBalance(address) function.

getColorCall

Function with signature getColor() and selector 0x9a86139b.

getColorReturn

Container type for the return parameters of the getColor() function.

getControllerCall

Function with signature getController() and selector 0x3018205f.

getControllerReturn

Container type for the return parameters of the getController() function.

getCurrentTokensCall

Function with signature getCurrentTokens() and selector 0xcc77828d.

getCurrentTokensReturn

Container type for the return parameters of the getCurrentTokens() function.

getDenormalizedWeightCall

Function with signature getDenormalizedWeight(address) and selector 0x948d8ce6.

getDenormalizedWeightReturn

Container type for the return parameters of the getDenormalizedWeight(address) function.

getFinalTokensCall

Function with signature getFinalTokens() and selector 0xbe3bbd2e.

getFinalTokensReturn

Container type for the return parameters of the getFinalTokens() function.

getNormalizedWeightCall

Function with signature getNormalizedWeight(address) and selector 0xf1b8a9b7.

getNormalizedWeightReturn

Container type for the return parameters of the getNormalizedWeight(address) function.

getNumTokensCall

Function with signature getNumTokens() and selector 0xcd2ed8fb.

getNumTokensReturn

Container type for the return parameters of the getNumTokens() function.

getSpotPriceCall

Function with signature getSpotPrice(address,address) and selector 0x15e84af9.

getSpotPriceReturn

Container type for the return parameters of the getSpotPrice(address,address) function.

getSpotPriceSansFeeCall

Function with signature getSpotPriceSansFee(address,address) and selector 0x1446a7ff.

getSpotPriceSansFeeReturn

Container type for the return parameters of the getSpotPriceSansFee(address,address) function.

getSwapFeeCall

Function with signature getSwapFee() and selector 0xd4cadf68.

getSwapFeeReturn

Container type for the return parameters of the getSwapFee() function.

getTotalDenormalizedWeightCall

Function with signature getTotalDenormalizedWeight() and selector 0x936c3477.

getTotalDenormalizedWeightReturn

Container type for the return parameters of the getTotalDenormalizedWeight() function.

gulpCall

Function with signature gulp(address) and selector 0x8c28cbe8.

gulpReturn

Container type for the return parameters of the gulp(address) function.

increaseApprovalCall

Function with signature increaseApproval(address,uint256) and selector 0xd73dd623.

increaseApprovalReturn

Container type for the return parameters of the increaseApproval(address,uint256) function.

isBoundCall

Function with signature isBound(address) and selector 0x2f37b624.

isBoundReturn

Container type for the return parameters of the isBound(address) function.

isFinalizedCall

Function with signature isFinalized() and selector 0x8d4e4083.

isFinalizedReturn

Container type for the return parameters of the isFinalized() function.

isPublicSwapCall

Function with signature isPublicSwap() and selector 0xfde924f7.

isPublicSwapReturn

Container type for the return parameters of the isPublicSwap() function.

joinPoolCall

Function with signature joinPool(uint256,uint256[]) and selector 0x4f69c0d4.

joinPoolReturn

Container type for the return parameters of the joinPool(uint256,uint256[]) function.

joinswapExternAmountInCall

Function with signature joinswapExternAmountIn(address,uint256,uint256) and selector 0x5db34277.

joinswapExternAmountInReturn

Container type for the return parameters of the joinswapExternAmountIn(address,uint256,uint256) function.

joinswapPoolAmountOutCall

Function with signature joinswapPoolAmountOut(address,uint256,uint256) and selector 0x6d06dfa0.

joinswapPoolAmountOutReturn

Container type for the return parameters of the joinswapPoolAmountOut(address,uint256,uint256) function.

nameCall

Function with signature name() and selector 0x06fdde03.

nameReturn

Container type for the return parameters of the name() function.

rebindCall

Function with signature rebind(address,uint256,uint256) and selector 0x3fdddaa2.

rebindReturn

Container type for the return parameters of the rebind(address,uint256,uint256) function.

setControllerCall

Function with signature setController(address) and selector 0x92eefe9b.

setControllerReturn

Container type for the return parameters of the setController(address) function.

setPublicSwapCall

Function with signature setPublicSwap(bool) and selector 0x49b59552.

setPublicSwapReturn

Container type for the return parameters of the setPublicSwap(bool) function.

setSwapFeeCall

Function with signature setSwapFee(uint256) and selector 0x34e19907.

setSwapFeeReturn

Container type for the return parameters of the setSwapFee(uint256) function.

swapExactAmountInCall

Function with signature swapExactAmountIn(address,uint256,address,uint256,uint256) and selector 0x8201aa3f.

swapExactAmountInReturn

Container type for the return parameters of the swapExactAmountIn(address,uint256,address,uint256,uint256) function.

swapExactAmountOutCall

Function with signature swapExactAmountOut(address,uint256,address,uint256,uint256) and selector 0x7c5e9ea4.

swapExactAmountOutReturn

Container type for the return parameters of the swapExactAmountOut(address,uint256,address,uint256,uint256) function.

symbolCall

Function with signature symbol() and selector 0x95d89b41.

symbolReturn

Container type for the return parameters of the symbol() function.

totalSupplyCall

Function with signature totalSupply() and selector 0x18160ddd.

totalSupplyReturn

Container type for the return parameters of the totalSupply() function.

transferCall

Function with signature transfer(address,uint256) and selector 0xa9059cbb.

transferFromCall

Function with signature transferFrom(address,address,uint256) and selector 0x23b872dd.

transferFromReturn

Container type for the return parameters of the transferFrom(address,address,uint256) function.

transferReturn

Container type for the return parameters of the transfer(address,uint256) function.

unbindCall

Function with signature unbind(address) and selector 0xcf5e7bd3.

unbindReturn

Container type for the return parameters of the unbind(address) function.

Enums

BalancerV1Calls

Container for all the BalancerV1 function calls.

BalancerV1Events

Container for all the BalancerV1 events.