Module brontes_classifier::CurveCryptoSwap

Generated by the following Solidity interface…

interface CurveCryptoSwap {
    event AddLiquidity(address indexed provider, uint256[2] token_amounts, uint256 fee, uint256 token_supply);
    event ClaimAdminFee(address indexed admin, uint256 tokens);
    event CommitNewParameters(uint256 indexed deadline, uint256 admin_fee, uint256 mid_fee, uint256 out_fee, uint256 fee_gamma, uint256 allowed_extra_profit, uint256 adjustment_step, uint256 ma_half_time);
    event NewParameters(uint256 admin_fee, uint256 mid_fee, uint256 out_fee, uint256 fee_gamma, uint256 allowed_extra_profit, uint256 adjustment_step, uint256 ma_half_time);
    event RampAgamma(uint256 initial_A, uint256 future_A, uint256 initial_gamma, uint256 future_gamma, uint256 initial_time, uint256 future_time);
    event RemoveLiquidity(address indexed provider, uint256[2] token_amounts, uint256 token_supply);
    event RemoveLiquidityOne(address indexed provider, uint256 token_amount, uint256 coin_index, uint256 coin_amount);
    event StopRampA(uint256 current_A, uint256 current_gamma, uint256 time);
    event TokenExchange(address indexed buyer, uint256 sold_id, uint256 tokens_sold, uint256 bought_id, uint256 tokens_bought);

    fallback() external payable;

    function A() external view returns (uint256);
    function D() external view returns (uint256);
    function add_liquidity(uint256[2] memory amounts, uint256 min_mint_amount) external payable returns (uint256);
    function add_liquidity(uint256[2] memory amounts, uint256 min_mint_amount, bool use_eth) external payable returns (uint256);
    function add_liquidity(uint256[2] memory amounts, uint256 min_mint_amount, bool use_eth, address receiver) external payable returns (uint256);
    function adjustment_step() external view returns (uint256);
    function admin_actions_deadline() external view returns (uint256);
    function admin_fee() external view returns (uint256);
    function allowed_extra_profit() external view returns (uint256);
    function apply_new_parameters() external;
    function balances(uint256 arg0) external view returns (uint256);
    function calc_token_amount(uint256[2] memory amounts) external view returns (uint256);
    function calc_withdraw_one_coin(uint256 token_amount, uint256 i) external view returns (uint256);
    function claim_admin_fees() external;
    function coins(uint256 arg0) external view returns (address);
    function commit_new_parameters(uint256 _new_mid_fee, uint256 _new_out_fee, uint256 _new_admin_fee, uint256 _new_fee_gamma, uint256 _new_allowed_extra_profit, uint256 _new_adjustment_step, uint256 _new_ma_half_time) external;
    function exchange(uint256 i, uint256 j, uint256 dx, uint256 min_dy) external payable returns (uint256);
    function exchange(uint256 i, uint256 j, uint256 dx, uint256 min_dy, bool use_eth) external payable returns (uint256);
    function exchange(uint256 i, uint256 j, uint256 dx, uint256 min_dy, bool use_eth, address receiver) external payable returns (uint256);
    function exchange_extended(uint256 i, uint256 j, uint256 dx, uint256 min_dy, bool use_eth, address sender, address receiver, bytes32 cb) external payable returns (uint256);
    function exchange_underlying(uint256 i, uint256 j, uint256 dx, uint256 min_dy) external payable returns (uint256);
    function exchange_underlying(uint256 i, uint256 j, uint256 dx, uint256 min_dy, address receiver) external payable returns (uint256);
    function factory() external view returns (address);
    function fee() external view returns (uint256);
    function fee_gamma() external view returns (uint256);
    function future_A_gamma() external view returns (uint256);
    function future_A_gamma_time() external view returns (uint256);
    function future_adjustment_step() external view returns (uint256);
    function future_admin_fee() external view returns (uint256);
    function future_allowed_extra_profit() external view returns (uint256);
    function future_fee_gamma() external view returns (uint256);
    function future_ma_half_time() external view returns (uint256);
    function future_mid_fee() external view returns (uint256);
    function future_out_fee() external view returns (uint256);
    function gamma() external view returns (uint256);
    function get_dy(uint256 i, uint256 j, uint256 dx) external view returns (uint256);
    function get_virtual_price() external view returns (uint256);
    function initial_A_gamma() external view returns (uint256);
    function initial_A_gamma_time() external view returns (uint256);
    function initialize(uint256 A, uint256 gamma, uint256 mid_fee, uint256 out_fee, uint256 allowed_extra_profit, uint256 fee_gamma, uint256 adjustment_step, uint256 admin_fee, uint256 ma_half_time, uint256 initial_price, address _token, address[2] memory _coins, uint256 _precisions) external;
    function last_prices() external view returns (uint256);
    function last_prices_timestamp() external view returns (uint256);
    function lp_price() external view returns (uint256);
    function ma_half_time() external view returns (uint256);
    function mid_fee() external view returns (uint256);
    function out_fee() external view returns (uint256);
    function price_oracle() external view returns (uint256);
    function price_scale() external view returns (uint256);
    function ramp_A_gamma(uint256 future_A, uint256 future_gamma, uint256 future_time) external;
    function remove_liquidity(uint256 _amount, uint256[2] memory min_amounts) external;
    function remove_liquidity(uint256 _amount, uint256[2] memory min_amounts, bool use_eth) external;
    function remove_liquidity(uint256 _amount, uint256[2] memory min_amounts, bool use_eth, address receiver) external;
    function remove_liquidity_one_coin(uint256 token_amount, uint256 i, uint256 min_amount) external returns (uint256);
    function remove_liquidity_one_coin(uint256 token_amount, uint256 i, uint256 min_amount, bool use_eth) external returns (uint256);
    function remove_liquidity_one_coin(uint256 token_amount, uint256 i, uint256 min_amount, bool use_eth, address receiver) external returns (uint256);
    function revert_new_parameters() external;
    function stop_ramp_A_gamma() external;
    function token() external view returns (address);
    function virtual_price() external view returns (uint256);
    function xcp_profit() external view returns (uint256);
    function xcp_profit_a() external view returns (uint256);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_weth",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "fallback",
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "A",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "D",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "add_liquidity",
    "inputs": [
      {
        "name": "amounts",
        "type": "uint256[2]"
      },
      {
        "name": "min_mint_amount",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "add_liquidity",
    "inputs": [
      {
        "name": "amounts",
        "type": "uint256[2]"
      },
      {
        "name": "min_mint_amount",
        "type": "uint256"
      },
      {
        "name": "use_eth",
        "type": "bool"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "add_liquidity",
    "inputs": [
      {
        "name": "amounts",
        "type": "uint256[2]"
      },
      {
        "name": "min_mint_amount",
        "type": "uint256"
      },
      {
        "name": "use_eth",
        "type": "bool"
      },
      {
        "name": "receiver",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "adjustment_step",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "admin_actions_deadline",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "admin_fee",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "allowed_extra_profit",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "apply_new_parameters",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "balances",
    "inputs": [
      {
        "name": "arg0",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "calc_token_amount",
    "inputs": [
      {
        "name": "amounts",
        "type": "uint256[2]"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "calc_withdraw_one_coin",
    "inputs": [
      {
        "name": "token_amount",
        "type": "uint256"
      },
      {
        "name": "i",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "claim_admin_fees",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "coins",
    "inputs": [
      {
        "name": "arg0",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "commit_new_parameters",
    "inputs": [
      {
        "name": "_new_mid_fee",
        "type": "uint256"
      },
      {
        "name": "_new_out_fee",
        "type": "uint256"
      },
      {
        "name": "_new_admin_fee",
        "type": "uint256"
      },
      {
        "name": "_new_fee_gamma",
        "type": "uint256"
      },
      {
        "name": "_new_allowed_extra_profit",
        "type": "uint256"
      },
      {
        "name": "_new_adjustment_step",
        "type": "uint256"
      },
      {
        "name": "_new_ma_half_time",
        "type": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "exchange",
    "inputs": [
      {
        "name": "i",
        "type": "uint256"
      },
      {
        "name": "j",
        "type": "uint256"
      },
      {
        "name": "dx",
        "type": "uint256"
      },
      {
        "name": "min_dy",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "exchange",
    "inputs": [
      {
        "name": "i",
        "type": "uint256"
      },
      {
        "name": "j",
        "type": "uint256"
      },
      {
        "name": "dx",
        "type": "uint256"
      },
      {
        "name": "min_dy",
        "type": "uint256"
      },
      {
        "name": "use_eth",
        "type": "bool"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "exchange",
    "inputs": [
      {
        "name": "i",
        "type": "uint256"
      },
      {
        "name": "j",
        "type": "uint256"
      },
      {
        "name": "dx",
        "type": "uint256"
      },
      {
        "name": "min_dy",
        "type": "uint256"
      },
      {
        "name": "use_eth",
        "type": "bool"
      },
      {
        "name": "receiver",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "exchange_extended",
    "inputs": [
      {
        "name": "i",
        "type": "uint256"
      },
      {
        "name": "j",
        "type": "uint256"
      },
      {
        "name": "dx",
        "type": "uint256"
      },
      {
        "name": "min_dy",
        "type": "uint256"
      },
      {
        "name": "use_eth",
        "type": "bool"
      },
      {
        "name": "sender",
        "type": "address"
      },
      {
        "name": "receiver",
        "type": "address"
      },
      {
        "name": "cb",
        "type": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "exchange_underlying",
    "inputs": [
      {
        "name": "i",
        "type": "uint256"
      },
      {
        "name": "j",
        "type": "uint256"
      },
      {
        "name": "dx",
        "type": "uint256"
      },
      {
        "name": "min_dy",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "exchange_underlying",
    "inputs": [
      {
        "name": "i",
        "type": "uint256"
      },
      {
        "name": "j",
        "type": "uint256"
      },
      {
        "name": "dx",
        "type": "uint256"
      },
      {
        "name": "min_dy",
        "type": "uint256"
      },
      {
        "name": "receiver",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "factory",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "fee",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "fee_gamma",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_A_gamma",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_A_gamma_time",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_adjustment_step",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_admin_fee",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_allowed_extra_profit",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_fee_gamma",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_ma_half_time",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_mid_fee",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_out_fee",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "gamma",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_dy",
    "inputs": [
      {
        "name": "i",
        "type": "uint256"
      },
      {
        "name": "j",
        "type": "uint256"
      },
      {
        "name": "dx",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_virtual_price",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "initial_A_gamma",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "initial_A_gamma_time",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "initialize",
    "inputs": [
      {
        "name": "A",
        "type": "uint256"
      },
      {
        "name": "gamma",
        "type": "uint256"
      },
      {
        "name": "mid_fee",
        "type": "uint256"
      },
      {
        "name": "out_fee",
        "type": "uint256"
      },
      {
        "name": "allowed_extra_profit",
        "type": "uint256"
      },
      {
        "name": "fee_gamma",
        "type": "uint256"
      },
      {
        "name": "adjustment_step",
        "type": "uint256"
      },
      {
        "name": "admin_fee",
        "type": "uint256"
      },
      {
        "name": "ma_half_time",
        "type": "uint256"
      },
      {
        "name": "initial_price",
        "type": "uint256"
      },
      {
        "name": "_token",
        "type": "address"
      },
      {
        "name": "_coins",
        "type": "address[2]"
      },
      {
        "name": "_precisions",
        "type": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "last_prices",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "last_prices_timestamp",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "lp_price",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "ma_half_time",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "mid_fee",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "out_fee",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "price_oracle",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "price_scale",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "ramp_A_gamma",
    "inputs": [
      {
        "name": "future_A",
        "type": "uint256"
      },
      {
        "name": "future_gamma",
        "type": "uint256"
      },
      {
        "name": "future_time",
        "type": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "remove_liquidity",
    "inputs": [
      {
        "name": "_amount",
        "type": "uint256"
      },
      {
        "name": "min_amounts",
        "type": "uint256[2]"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "remove_liquidity",
    "inputs": [
      {
        "name": "_amount",
        "type": "uint256"
      },
      {
        "name": "min_amounts",
        "type": "uint256[2]"
      },
      {
        "name": "use_eth",
        "type": "bool"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "remove_liquidity",
    "inputs": [
      {
        "name": "_amount",
        "type": "uint256"
      },
      {
        "name": "min_amounts",
        "type": "uint256[2]"
      },
      {
        "name": "use_eth",
        "type": "bool"
      },
      {
        "name": "receiver",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "remove_liquidity_one_coin",
    "inputs": [
      {
        "name": "token_amount",
        "type": "uint256"
      },
      {
        "name": "i",
        "type": "uint256"
      },
      {
        "name": "min_amount",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "remove_liquidity_one_coin",
    "inputs": [
      {
        "name": "token_amount",
        "type": "uint256"
      },
      {
        "name": "i",
        "type": "uint256"
      },
      {
        "name": "min_amount",
        "type": "uint256"
      },
      {
        "name": "use_eth",
        "type": "bool"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "remove_liquidity_one_coin",
    "inputs": [
      {
        "name": "token_amount",
        "type": "uint256"
      },
      {
        "name": "i",
        "type": "uint256"
      },
      {
        "name": "min_amount",
        "type": "uint256"
      },
      {
        "name": "use_eth",
        "type": "bool"
      },
      {
        "name": "receiver",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "revert_new_parameters",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "stop_ramp_A_gamma",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "token",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "virtual_price",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "xcp_profit",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "xcp_profit_a",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "event",
    "name": "AddLiquidity",
    "inputs": [
      {
        "name": "provider",
        "type": "address",
        "indexed": true
      },
      {
        "name": "token_amounts",
        "type": "uint256[2]",
        "indexed": false
      },
      {
        "name": "fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "token_supply",
        "type": "uint256",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ClaimAdminFee",
    "inputs": [
      {
        "name": "admin",
        "type": "address",
        "indexed": true
      },
      {
        "name": "tokens",
        "type": "uint256",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "CommitNewParameters",
    "inputs": [
      {
        "name": "deadline",
        "type": "uint256",
        "indexed": true
      },
      {
        "name": "admin_fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "mid_fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "out_fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "fee_gamma",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "allowed_extra_profit",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "adjustment_step",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "ma_half_time",
        "type": "uint256",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NewParameters",
    "inputs": [
      {
        "name": "admin_fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "mid_fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "out_fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "fee_gamma",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "allowed_extra_profit",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "adjustment_step",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "ma_half_time",
        "type": "uint256",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RampAgamma",
    "inputs": [
      {
        "name": "initial_A",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "future_A",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "initial_gamma",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "future_gamma",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "initial_time",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "future_time",
        "type": "uint256",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RemoveLiquidity",
    "inputs": [
      {
        "name": "provider",
        "type": "address",
        "indexed": true
      },
      {
        "name": "token_amounts",
        "type": "uint256[2]",
        "indexed": false
      },
      {
        "name": "token_supply",
        "type": "uint256",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RemoveLiquidityOne",
    "inputs": [
      {
        "name": "provider",
        "type": "address",
        "indexed": true
      },
      {
        "name": "token_amount",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "coin_index",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "coin_amount",
        "type": "uint256",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "StopRampA",
    "inputs": [
      {
        "name": "current_A",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "current_gamma",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "time",
        "type": "uint256",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "TokenExchange",
    "inputs": [
      {
        "name": "buyer",
        "type": "address",
        "indexed": true
      },
      {
        "name": "sold_id",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "tokens_sold",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "bought_id",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "tokens_bought",
        "type": "uint256",
        "indexed": false
      }
    ],
    "anonymous": false
  }
]

Structs

ACall

Function with signature A() and selector 0xf446c1d0.

AReturn

Container type for the return parameters of the A() function.

AddLiquidity

Event with signature AddLiquidity(address,uint256[2],uint256,uint256) and selector 0x540ab385f9b5d450a27404172caade516b3ba3f4be88239ac56a2ad1de2a1f5a.

ClaimAdminFee

Event with signature ClaimAdminFee(address,uint256) and selector 0x6059a38198b1dc42b3791087d1ff0fbd72b3179553c25f678cd246f52ffaaf59.

CommitNewParameters

Event with signature CommitNewParameters(uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256) and selector 0x913fde9a37e1f8ab67876a4d0ce80790d764fcfc5692f4529526df9c6bdde553.

DCall

Function with signature D() and selector 0x0f529ba2.

DReturn

Container type for the return parameters of the D() function.

NewParameters

Event with signature NewParameters(uint256,uint256,uint256,uint256,uint256,uint256,uint256) and selector 0x1c65bbdc939f346e5d6f0bde1f072819947438d4fc7b182cc59c2f6dc5504087.

RampAgamma

Event with signature RampAgamma(uint256,uint256,uint256,uint256,uint256,uint256) and selector 0xe35f0559b0642164e286b30df2077ec3a05426617a25db7578fd20ba39a6cd05.

RemoveLiquidity

Event with signature RemoveLiquidity(address,uint256[2],uint256) and selector 0xdd3c0336a16f1b64f172b7bb0dad5b2b3c7c76f91e8c4aafd6aae60dce800153.

RemoveLiquidityOne

Event with signature RemoveLiquidityOne(address,uint256,uint256,uint256) and selector 0x5ad056f2e28a8cec232015406b843668c1e36cda598127ec3b8c59b8c72773a0.

StopRampA

Event with signature StopRampA(uint256,uint256,uint256) and selector 0x5f0e7fba3d100c9e19446e1c92fe436f0a9a22fe99669360e4fdd6d3de2fc284.

TokenExchange

Event with signature TokenExchange(address,uint256,uint256,uint256,uint256) and selector 0xb2e76ae99761dc136e598d4a629bb347eccb9532a5f8bbd72e18467c3c34cc98.

add_liquidity_0Call

Function with signature add_liquidity(uint256[2],uint256) and selector 0x0b4c7e4d.

add_liquidity_0Return

Container type for the return parameters of the add_liquidity(uint256[2],uint256) function.

add_liquidity_1Call

Function with signature add_liquidity(uint256[2],uint256,bool) and selector 0xee22be23.

add_liquidity_1Return

Container type for the return parameters of the add_liquidity(uint256[2],uint256,bool) function.

add_liquidity_2Call

Function with signature add_liquidity(uint256[2],uint256,bool,address) and selector 0x7328333b.

add_liquidity_2Return

Container type for the return parameters of the add_liquidity(uint256[2],uint256,bool,address) function.

adjustment_stepCall

Function with signature adjustment_step() and selector 0x083812e5.

adjustment_stepReturn

Container type for the return parameters of the adjustment_step() function.

admin_actions_deadlineCall

Function with signature admin_actions_deadline() and selector 0x405e28f8.

admin_actions_deadlineReturn

Container type for the return parameters of the admin_actions_deadline() function.

admin_feeCall

Function with signature admin_fee() and selector 0xfee3f7f9.

admin_feeReturn

Container type for the return parameters of the admin_fee() function.

allowed_extra_profitCall

Function with signature allowed_extra_profit() and selector 0x49fe9e77.

allowed_extra_profitReturn

Container type for the return parameters of the allowed_extra_profit() function.

apply_new_parametersCall

Function with signature apply_new_parameters() and selector 0x2a7dd7cd.

apply_new_parametersReturn

Container type for the return parameters of the apply_new_parameters() function.

balancesCall

Function with signature balances(uint256) and selector 0x4903b0d1.

balancesReturn

Container type for the return parameters of the balances(uint256) function.

calc_token_amountCall

Function with signature calc_token_amount(uint256[2]) and selector 0x8d8ea727.

calc_token_amountReturn

Container type for the return parameters of the calc_token_amount(uint256[2]) function.

calc_withdraw_one_coinCall

Function with signature calc_withdraw_one_coin(uint256,uint256) and selector 0x4fb08c5e.

calc_withdraw_one_coinReturn

Container type for the return parameters of the calc_withdraw_one_coin(uint256,uint256) function.

claim_admin_feesCall

Function with signature claim_admin_fees() and selector 0xc93f49e8.

claim_admin_feesReturn

Container type for the return parameters of the claim_admin_fees() function.

coinsCall

Function with signature coins(uint256) and selector 0xc6610657.

coinsReturn

Container type for the return parameters of the coins(uint256) function.

commit_new_parametersCall

Function with signature commit_new_parameters(uint256,uint256,uint256,uint256,uint256,uint256,uint256) and selector 0xa43c3351.

commit_new_parametersReturn

Container type for the return parameters of the commit_new_parameters(uint256,uint256,uint256,uint256,uint256,uint256,uint256) function.

exchange_0Call

Function with signature exchange(uint256,uint256,uint256,uint256) and selector 0x5b41b908.

exchange_0Return

Container type for the return parameters of the exchange(uint256,uint256,uint256,uint256) function.

exchange_1Call

Function with signature exchange(uint256,uint256,uint256,uint256,bool) and selector 0x394747c5.

exchange_1Return

Container type for the return parameters of the exchange(uint256,uint256,uint256,uint256,bool) function.

exchange_2Call

Function with signature exchange(uint256,uint256,uint256,uint256,bool,address) and selector 0xce7d6503.

exchange_2Return

Container type for the return parameters of the exchange(uint256,uint256,uint256,uint256,bool,address) function.

exchange_extendedCall

Function with signature exchange_extended(uint256,uint256,uint256,uint256,bool,address,address,bytes32) and selector 0xdd96994f.

exchange_extendedReturn

Container type for the return parameters of the exchange_extended(uint256,uint256,uint256,uint256,bool,address,address,bytes32) function.

exchange_underlying_0Call

Function with signature exchange_underlying(uint256,uint256,uint256,uint256) and selector 0x65b2489b.

exchange_underlying_0Return

Container type for the return parameters of the exchange_underlying(uint256,uint256,uint256,uint256) function.

exchange_underlying_1Call

Function with signature exchange_underlying(uint256,uint256,uint256,uint256,address) and selector 0xe2ad025a.

exchange_underlying_1Return

Container type for the return parameters of the exchange_underlying(uint256,uint256,uint256,uint256,address) function.

factoryCall

Function with signature factory() and selector 0xc45a0155.

factoryReturn

Container type for the return parameters of the factory() function.

feeCall

Function with signature fee() and selector 0xddca3f43.

feeReturn

Container type for the return parameters of the fee() function.

fee_gammaCall

Function with signature fee_gamma() and selector 0x72d4f0e2.

fee_gammaReturn

Container type for the return parameters of the fee_gamma() function.

future_A_gammaCall

Function with signature future_A_gamma() and selector 0xf30cfad5.

future_A_gammaReturn

Container type for the return parameters of the future_A_gamma() function.

future_A_gamma_timeCall

Function with signature future_A_gamma_time() and selector 0xf9ed9597.

future_A_gamma_timeReturn

Container type for the return parameters of the future_A_gamma_time() function.

future_adjustment_stepCall

Function with signature future_adjustment_step() and selector 0x4ea12c7d.

future_adjustment_stepReturn

Container type for the return parameters of the future_adjustment_step() function.

future_admin_feeCall

Function with signature future_admin_fee() and selector 0xe3824462.

future_admin_feeReturn

Container type for the return parameters of the future_admin_fee() function.

future_allowed_extra_profitCall

Function with signature future_allowed_extra_profit() and selector 0x727ced57.

future_allowed_extra_profitReturn

Container type for the return parameters of the future_allowed_extra_profit() function.

future_fee_gammaCall

Function with signature future_fee_gamma() and selector 0xd7c3dcbe.

future_fee_gammaReturn

Container type for the return parameters of the future_fee_gamma() function.

future_ma_half_timeCall

Function with signature future_ma_half_time() and selector 0x0c5e23d4.

future_ma_half_timeReturn

Container type for the return parameters of the future_ma_half_time() function.

future_mid_feeCall

Function with signature future_mid_fee() and selector 0x7cf9aedc.

future_mid_feeReturn

Container type for the return parameters of the future_mid_fee() function.

future_out_feeCall

Function with signature future_out_fee() and selector 0x7d1b060c.

future_out_feeReturn

Container type for the return parameters of the future_out_fee() function.

gammaCall

Function with signature gamma() and selector 0xb1373929.

gammaReturn

Container type for the return parameters of the gamma() function.

get_dyCall

Function with signature get_dy(uint256,uint256,uint256) and selector 0x556d6e9f.

get_dyReturn

Container type for the return parameters of the get_dy(uint256,uint256,uint256) function.

get_virtual_priceCall

Function with signature get_virtual_price() and selector 0xbb7b8b80.

get_virtual_priceReturn

Container type for the return parameters of the get_virtual_price() function.

initial_A_gammaCall

Function with signature initial_A_gamma() and selector 0x204fe3d5.

initial_A_gammaReturn

Container type for the return parameters of the initial_A_gamma() function.

initial_A_gamma_timeCall

Function with signature initial_A_gamma_time() and selector 0xe89876ff.

initial_A_gamma_timeReturn

Container type for the return parameters of the initial_A_gamma_time() function.

initializeCall

Function with signature initialize(uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,address,address[2],uint256) and selector 0xa39e95c5.

initializeReturn

Container type for the return parameters of the initialize(uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,address,address[2],uint256) function.

last_pricesCall

Function with signature last_prices() and selector 0xc146bf94.

last_pricesReturn

Container type for the return parameters of the last_prices() function.

last_prices_timestampCall

Function with signature last_prices_timestamp() and selector 0x6112c747.

last_prices_timestampReturn

Container type for the return parameters of the last_prices_timestamp() function.

lp_priceCall

Function with signature lp_price() and selector 0x54f0f7d5.

lp_priceReturn

Container type for the return parameters of the lp_price() function.

ma_half_timeCall

Function with signature ma_half_time() and selector 0x662b6274.

ma_half_timeReturn

Container type for the return parameters of the ma_half_time() function.

mid_feeCall

Function with signature mid_fee() and selector 0x92526c0c.

mid_feeReturn

Container type for the return parameters of the mid_fee() function.

out_feeCall

Function with signature out_fee() and selector 0xee8de675.

out_feeReturn

Container type for the return parameters of the out_fee() function.

price_oracleCall

Function with signature price_oracle() and selector 0x86fc88d3.

price_oracleReturn

Container type for the return parameters of the price_oracle() function.

price_scaleCall

Function with signature price_scale() and selector 0xb9e8c9fd.

price_scaleReturn

Container type for the return parameters of the price_scale() function.

ramp_A_gammaCall

Function with signature ramp_A_gamma(uint256,uint256,uint256) and selector 0x5e248072.

ramp_A_gammaReturn

Container type for the return parameters of the ramp_A_gamma(uint256,uint256,uint256) function.

remove_liquidity_0Call

Function with signature remove_liquidity(uint256,uint256[2]) and selector 0x5b36389c.

remove_liquidity_0Return

Container type for the return parameters of the remove_liquidity(uint256,uint256[2]) function.

remove_liquidity_1Call

Function with signature remove_liquidity(uint256,uint256[2],bool) and selector 0x269b5581.

remove_liquidity_1Return

Container type for the return parameters of the remove_liquidity(uint256,uint256[2],bool) function.

remove_liquidity_2Call

Function with signature remove_liquidity(uint256,uint256[2],bool,address) and selector 0x1808e84a.

remove_liquidity_2Return

Container type for the return parameters of the remove_liquidity(uint256,uint256[2],bool,address) function.

remove_liquidity_one_coin_0Call

Function with signature remove_liquidity_one_coin(uint256,uint256,uint256) and selector 0xf1dc3cc9.

remove_liquidity_one_coin_0Return

Container type for the return parameters of the remove_liquidity_one_coin(uint256,uint256,uint256) function.

remove_liquidity_one_coin_1Call

Function with signature remove_liquidity_one_coin(uint256,uint256,uint256,bool) and selector 0x8f15b6b5.

remove_liquidity_one_coin_1Return

Container type for the return parameters of the remove_liquidity_one_coin(uint256,uint256,uint256,bool) function.

remove_liquidity_one_coin_2Call

Function with signature remove_liquidity_one_coin(uint256,uint256,uint256,bool,address) and selector 0x07329bcd.

remove_liquidity_one_coin_2Return

Container type for the return parameters of the remove_liquidity_one_coin(uint256,uint256,uint256,bool,address) function.

revert_new_parametersCall

Function with signature revert_new_parameters() and selector 0x226840fb.

revert_new_parametersReturn

Container type for the return parameters of the revert_new_parameters() function.

stop_ramp_A_gammaCall

Function with signature stop_ramp_A_gamma() and selector 0x244c7c2e.

stop_ramp_A_gammaReturn

Container type for the return parameters of the stop_ramp_A_gamma() function.

tokenCall

Function with signature token() and selector 0xfc0c546a.

tokenReturn

Container type for the return parameters of the token() function.

virtual_priceCall

Function with signature virtual_price() and selector 0x0c46b72a.

virtual_priceReturn

Container type for the return parameters of the virtual_price() function.

xcp_profitCall

Function with signature xcp_profit() and selector 0x7ba1a74d.

xcp_profitReturn

Container type for the return parameters of the xcp_profit() function.

xcp_profit_aCall

Function with signature xcp_profit_a() and selector 0x0b7b594b.

xcp_profit_aReturn

Container type for the return parameters of the xcp_profit_a() function.

Enums

CurveCryptoSwapCalls

Container for all the CurveCryptoSwap function calls.

CurveCryptoSwapEvents

Container for all the CurveCryptoSwap events.