Enum brontes_classifier::CurveCryptoSwap::CurveCryptoSwapCalls

pub enum CurveCryptoSwapCalls {

    A(ACall),
    D(DCall),
    add_liquidity_0(add_liquidity_0Call),
    add_liquidity_1(add_liquidity_1Call),
    add_liquidity_2(add_liquidity_2Call),
    adjustment_step(adjustment_stepCall),
    admin_actions_deadline(admin_actions_deadlineCall),
    admin_fee(admin_feeCall),
    allowed_extra_profit(allowed_extra_profitCall),
    apply_new_parameters(apply_new_parametersCall),
    balances(balancesCall),
    calc_token_amount(calc_token_amountCall),
    calc_withdraw_one_coin(calc_withdraw_one_coinCall),
    claim_admin_fees(claim_admin_feesCall),
    coins(coinsCall),
    commit_new_parameters(commit_new_parametersCall),
    exchange_0(exchange_0Call),
    exchange_1(exchange_1Call),
    exchange_2(exchange_2Call),
    exchange_extended(exchange_extendedCall),
    exchange_underlying_0(exchange_underlying_0Call),
    exchange_underlying_1(exchange_underlying_1Call),
    factory(factoryCall),
    fee(feeCall),
    fee_gamma(fee_gammaCall),
    future_A_gamma(future_A_gammaCall),
    future_A_gamma_time(future_A_gamma_timeCall),
    future_adjustment_step(future_adjustment_stepCall),
    future_admin_fee(future_admin_feeCall),
    future_allowed_extra_profit(future_allowed_extra_profitCall),
    future_fee_gamma(future_fee_gammaCall),
    future_ma_half_time(future_ma_half_timeCall),
    future_mid_fee(future_mid_feeCall),
    future_out_fee(future_out_feeCall),
    gamma(gammaCall),
    get_dy(get_dyCall),
    get_virtual_price(get_virtual_priceCall),
    initial_A_gamma(initial_A_gammaCall),
    initial_A_gamma_time(initial_A_gamma_timeCall),
    initialize(initializeCall),
    last_prices(last_pricesCall),
    last_prices_timestamp(last_prices_timestampCall),
    lp_price(lp_priceCall),
    ma_half_time(ma_half_timeCall),
    mid_fee(mid_feeCall),
    out_fee(out_feeCall),
    price_oracle(price_oracleCall),
    price_scale(price_scaleCall),
    ramp_A_gamma(ramp_A_gammaCall),
    remove_liquidity_0(remove_liquidity_0Call),
    remove_liquidity_1(remove_liquidity_1Call),
    remove_liquidity_2(remove_liquidity_2Call),
    remove_liquidity_one_coin_0(remove_liquidity_one_coin_0Call),
    remove_liquidity_one_coin_1(remove_liquidity_one_coin_1Call),
    remove_liquidity_one_coin_2(remove_liquidity_one_coin_2Call),
    revert_new_parameters(revert_new_parametersCall),
    stop_ramp_A_gamma(stop_ramp_A_gammaCall),
    token(tokenCall),
    virtual_price(virtual_priceCall),
    xcp_profit(xcp_profitCall),
    xcp_profit_a(xcp_profit_aCall),

}

Container for all the CurveCryptoSwap function calls.

Variants

A(ACall)

D(DCall)

add_liquidity_0(add_liquidity_0Call)

add_liquidity_1(add_liquidity_1Call)

add_liquidity_2(add_liquidity_2Call)

adjustment_step(adjustment_stepCall)

admin_actions_deadline(admin_actions_deadlineCall)

admin_fee(admin_feeCall)

allowed_extra_profit(allowed_extra_profitCall)

apply_new_parameters(apply_new_parametersCall)

balances(balancesCall)

calc_token_amount(calc_token_amountCall)

calc_withdraw_one_coin(calc_withdraw_one_coinCall)

claim_admin_fees(claim_admin_feesCall)

coins(coinsCall)

commit_new_parameters(commit_new_parametersCall)

exchange_0(exchange_0Call)

exchange_1(exchange_1Call)

exchange_2(exchange_2Call)

exchange_extended(exchange_extendedCall)

exchange_underlying_0(exchange_underlying_0Call)

exchange_underlying_1(exchange_underlying_1Call)

factory(factoryCall)

fee(feeCall)

fee_gamma(fee_gammaCall)

future_A_gamma(future_A_gammaCall)

future_A_gamma_time(future_A_gamma_timeCall)

future_adjustment_step(future_adjustment_stepCall)

future_admin_fee(future_admin_feeCall)

future_allowed_extra_profit(future_allowed_extra_profitCall)

future_fee_gamma(future_fee_gammaCall)

future_ma_half_time(future_ma_half_timeCall)

future_mid_fee(future_mid_feeCall)

future_out_fee(future_out_feeCall)

gamma(gammaCall)

get_dy(get_dyCall)

get_virtual_price(get_virtual_priceCall)

initial_A_gamma(initial_A_gammaCall)

initial_A_gamma_time(initial_A_gamma_timeCall)

initialize(initializeCall)

last_prices(last_pricesCall)

last_prices_timestamp(last_prices_timestampCall)

lp_price(lp_priceCall)

ma_half_time(ma_half_timeCall)

mid_fee(mid_feeCall)

out_fee(out_feeCall)

price_oracle(price_oracleCall)

price_scale(price_scaleCall)

ramp_A_gamma(ramp_A_gammaCall)

remove_liquidity_0(remove_liquidity_0Call)

remove_liquidity_1(remove_liquidity_1Call)

remove_liquidity_2(remove_liquidity_2Call)

remove_liquidity_one_coin_0(remove_liquidity_one_coin_0Call)

remove_liquidity_one_coin_1(remove_liquidity_one_coin_1Call)

remove_liquidity_one_coin_2(remove_liquidity_one_coin_2Call)

revert_new_parameters(revert_new_parametersCall)

stop_ramp_A_gamma(stop_ramp_A_gammaCall)

token(tokenCall)

virtual_price(virtual_priceCall)

xcp_profit(xcp_profitCall)

xcp_profit_a(xcp_profit_aCall)

Implementations

impl CurveCryptoSwapCalls

pub const SELECTORS: &'static [[u8; 4]] = _

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants, as they are sorted instead of ordered by definition.