Module brontes_classifier::CurveCryptoSwapFactory

Generated by the following Solidity interface…

interface CurveCryptoSwapFactory {
    event CryptoPoolDeployed(address token, address[2] coins, uint256 A, uint256 gamma, uint256 mid_fee, uint256 out_fee, uint256 allowed_extra_profit, uint256 fee_gamma, uint256 adjustment_step, uint256 admin_fee, uint256 ma_half_time, uint256 initial_price, address deployer);
    event LiquidityGaugeDeployed(address pool, address token, address gauge);
    event TransferOwnership(address _old_owner, address _new_owner);
    event UpdateFeeReceiver(address _old_fee_receiver, address _new_fee_receiver);
    event UpdateGaugeImplementation(address _old_gauge_implementation, address _new_gauge_implementation);
    event UpdatePoolImplementation(address _old_pool_implementation, address _new_pool_implementation);
    event UpdateTokenImplementation(address _old_token_implementation, address _new_token_implementation);

    function accept_transfer_ownership() external;
    function admin() external view returns (address);
    function commit_transfer_ownership(address _addr) external;
    function deploy_gauge(address _pool) external returns (address);
    function deploy_pool(string memory _name, string memory _symbol, address[2] memory _coins, uint256 A, uint256 gamma, uint256 mid_fee, uint256 out_fee, uint256 allowed_extra_profit, uint256 fee_gamma, uint256 adjustment_step, uint256 admin_fee, uint256 ma_half_time, uint256 initial_price) external returns (address);
    function fee_receiver() external view returns (address);
    function find_pool_for_coins(address _from, address _to) external view returns (address);
    function find_pool_for_coins(address _from, address _to, uint256 i) external view returns (address);
    function future_admin() external view returns (address);
    function gauge_implementation() external view returns (address);
    function get_balances(address _pool) external view returns (uint256[2] memory);
    function get_coin_indices(address _pool, address _from, address _to) external view returns (uint256, uint256);
    function get_coins(address _pool) external view returns (address[2] memory);
    function get_decimals(address _pool) external view returns (uint256[2] memory);
    function get_eth_index(address _pool) external view returns (uint256);
    function get_gauge(address _pool) external view returns (address);
    function get_token(address _pool) external view returns (address);
    function pool_count() external view returns (uint256);
    function pool_implementation() external view returns (address);
    function pool_list(uint256 arg0) external view returns (address);
    function set_fee_receiver(address _fee_receiver) external;
    function set_gauge_implementation(address _gauge_implementation) external;
    function set_pool_implementation(address _pool_implementation) external;
    function set_token_implementation(address _token_implementation) external;
    function token_implementation() external view returns (address);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_fee_receiver",
        "type": "address"
      },
      {
        "name": "_pool_implementation",
        "type": "address"
      },
      {
        "name": "_token_implementation",
        "type": "address"
      },
      {
        "name": "_gauge_implementation",
        "type": "address"
      },
      {
        "name": "_weth",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "accept_transfer_ownership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "admin",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "commit_transfer_ownership",
    "inputs": [
      {
        "name": "_addr",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deploy_gauge",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deploy_pool",
    "inputs": [
      {
        "name": "_name",
        "type": "string"
      },
      {
        "name": "_symbol",
        "type": "string"
      },
      {
        "name": "_coins",
        "type": "address[2]"
      },
      {
        "name": "A",
        "type": "uint256"
      },
      {
        "name": "gamma",
        "type": "uint256"
      },
      {
        "name": "mid_fee",
        "type": "uint256"
      },
      {
        "name": "out_fee",
        "type": "uint256"
      },
      {
        "name": "allowed_extra_profit",
        "type": "uint256"
      },
      {
        "name": "fee_gamma",
        "type": "uint256"
      },
      {
        "name": "adjustment_step",
        "type": "uint256"
      },
      {
        "name": "admin_fee",
        "type": "uint256"
      },
      {
        "name": "ma_half_time",
        "type": "uint256"
      },
      {
        "name": "initial_price",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "fee_receiver",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "find_pool_for_coins",
    "inputs": [
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "find_pool_for_coins",
    "inputs": [
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      },
      {
        "name": "i",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_admin",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "gauge_implementation",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_balances",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[2]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_coin_indices",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      },
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      },
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_coins",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address[2]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_decimals",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[2]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_eth_index",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_gauge",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_token",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pool_count",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pool_implementation",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pool_list",
    "inputs": [
      {
        "name": "arg0",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "set_fee_receiver",
    "inputs": [
      {
        "name": "_fee_receiver",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "set_gauge_implementation",
    "inputs": [
      {
        "name": "_gauge_implementation",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "set_pool_implementation",
    "inputs": [
      {
        "name": "_pool_implementation",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "set_token_implementation",
    "inputs": [
      {
        "name": "_token_implementation",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "token_implementation",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "event",
    "name": "CryptoPoolDeployed",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "indexed": false
      },
      {
        "name": "coins",
        "type": "address[2]",
        "indexed": false
      },
      {
        "name": "A",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "gamma",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "mid_fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "out_fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "allowed_extra_profit",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "fee_gamma",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "adjustment_step",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "admin_fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "ma_half_time",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "initial_price",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "deployer",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "LiquidityGaugeDeployed",
    "inputs": [
      {
        "name": "pool",
        "type": "address",
        "indexed": false
      },
      {
        "name": "token",
        "type": "address",
        "indexed": false
      },
      {
        "name": "gauge",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "TransferOwnership",
    "inputs": [
      {
        "name": "_old_owner",
        "type": "address",
        "indexed": false
      },
      {
        "name": "_new_owner",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "UpdateFeeReceiver",
    "inputs": [
      {
        "name": "_old_fee_receiver",
        "type": "address",
        "indexed": false
      },
      {
        "name": "_new_fee_receiver",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "UpdateGaugeImplementation",
    "inputs": [
      {
        "name": "_old_gauge_implementation",
        "type": "address",
        "indexed": false
      },
      {
        "name": "_new_gauge_implementation",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "UpdatePoolImplementation",
    "inputs": [
      {
        "name": "_old_pool_implementation",
        "type": "address",
        "indexed": false
      },
      {
        "name": "_new_pool_implementation",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "UpdateTokenImplementation",
    "inputs": [
      {
        "name": "_old_token_implementation",
        "type": "address",
        "indexed": false
      },
      {
        "name": "_new_token_implementation",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  }
]

Structs

CryptoPoolDeployed

Event with signature CryptoPoolDeployed(address,address[2],uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,address) and selector 0x0394cb40d7dbe28dad1d4ee890bdd35bbb0d89e17924a80a542535e83d54ba14.

LiquidityGaugeDeployed

Event with signature LiquidityGaugeDeployed(address,address,address) and selector 0x1d6247eae69b5feb96b30be78552f35de45f61fdb6d6d7e1b08aae159b6226af.

TransferOwnership

Event with signature TransferOwnership(address,address) and selector 0x5c486528ec3e3f0ea91181cff8116f02bfa350e03b8b6f12e00765adbb5af85c.

UpdateFeeReceiver

Event with signature UpdateFeeReceiver(address,address) and selector 0x2861448678f0be67f11bfb5481b3e3b4cfeb3acc6126ad60a05f95bfc6530666.

UpdateGaugeImplementation

Event with signature UpdateGaugeImplementation(address,address) and selector 0x1fd705f9c77053962a503f2f2f57f0862b4c3af687c25615c13817a86946c359.

UpdatePoolImplementation

Event with signature UpdatePoolImplementation(address,address) and selector 0x0617fd31aa5ab95ec80eefc1eb61a2c477aa419d1d761b4e46f5f077e47852aa.

UpdateTokenImplementation

Event with signature UpdateTokenImplementation(address,address) and selector 0x1cc4f8e20b0cd3e5109eb156cadcfd3a5496ac0794c6085ca02afd7d710dd566.

accept_transfer_ownershipCall

Function with signature accept_transfer_ownership() and selector 0xe5ea47b8.

accept_transfer_ownershipReturn

Container type for the return parameters of the accept_transfer_ownership() function.

adminCall

Function with signature admin() and selector 0xf851a440.

adminReturn

Container type for the return parameters of the admin() function.

commit_transfer_ownershipCall

Function with signature commit_transfer_ownership(address) and selector 0x6b441a40.

commit_transfer_ownershipReturn

Container type for the return parameters of the commit_transfer_ownership(address) function.

deploy_gaugeCall

Function with signature deploy_gauge(address) and selector 0x96bebb34.

deploy_gaugeReturn

Container type for the return parameters of the deploy_gauge(address) function.

deploy_poolCall

Function with signature deploy_pool(string,string,address[2],uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256) and selector 0xc955fa04.

deploy_poolReturn

Container type for the return parameters of the deploy_pool(string,string,address[2],uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256) function.

fee_receiverCall

Function with signature fee_receiver() and selector 0xcab4d3db.

fee_receiverReturn

Container type for the return parameters of the fee_receiver() function.

find_pool_for_coins_0Call

Function with signature find_pool_for_coins(address,address) and selector 0xa87df06c.

find_pool_for_coins_0Return

Container type for the return parameters of the find_pool_for_coins(address,address) function.

find_pool_for_coins_1Call

Function with signature find_pool_for_coins(address,address,uint256) and selector 0x6982eb0b.

find_pool_for_coins_1Return

Container type for the return parameters of the find_pool_for_coins(address,address,uint256) function.

future_adminCall

Function with signature future_admin() and selector 0x17f7182a.

future_adminReturn

Container type for the return parameters of the future_admin() function.

gauge_implementationCall

Function with signature gauge_implementation() and selector 0x8df24207.

gauge_implementationReturn

Container type for the return parameters of the gauge_implementation() function.

get_balancesCall

Function with signature get_balances(address) and selector 0x92e3cc2d.

get_balancesReturn

Container type for the return parameters of the get_balances(address) function.

get_coin_indicesCall

Function with signature get_coin_indices(address,address,address) and selector 0xeb85226d.

get_coin_indicesReturn

Container type for the return parameters of the get_coin_indices(address,address,address) function.

get_coinsCall

Function with signature get_coins(address) and selector 0x9ac90d3d.

get_coinsReturn

Container type for the return parameters of the get_coins(address) function.

get_decimalsCall

Function with signature get_decimals(address) and selector 0x52b51555.

get_decimalsReturn

Container type for the return parameters of the get_decimals(address) function.

get_eth_indexCall

Function with signature get_eth_index(address) and selector 0xccb15605.

get_eth_indexReturn

Container type for the return parameters of the get_eth_index(address) function.

get_gaugeCall

Function with signature get_gauge(address) and selector 0xdaf297b9.

get_gaugeReturn

Container type for the return parameters of the get_gauge(address) function.

get_tokenCall

Function with signature get_token(address) and selector 0x977d9122.

get_tokenReturn

Container type for the return parameters of the get_token(address) function.

pool_countCall

Function with signature pool_count() and selector 0x956aae3a.

pool_countReturn

Container type for the return parameters of the pool_count() function.

pool_implementationCall

Function with signature pool_implementation() and selector 0x2489a2c3.

pool_implementationReturn

Container type for the return parameters of the pool_implementation() function.

pool_listCall

Function with signature pool_list(uint256) and selector 0x3a1d5d8e.

pool_listReturn

Container type for the return parameters of the pool_list(uint256) function.

set_fee_receiverCall

Function with signature set_fee_receiver(address) and selector 0xe41ab771.

set_fee_receiverReturn

Container type for the return parameters of the set_fee_receiver(address) function.

set_gauge_implementationCall

Function with signature set_gauge_implementation(address) and selector 0x8f03182c.

set_gauge_implementationReturn

Container type for the return parameters of the set_gauge_implementation(address) function.

set_pool_implementationCall

Function with signature set_pool_implementation(address) and selector 0x9ed796d0.

set_pool_implementationReturn

Container type for the return parameters of the set_pool_implementation(address) function.

set_token_implementationCall

Function with signature set_token_implementation(address) and selector 0x653023c2.

set_token_implementationReturn

Container type for the return parameters of the set_token_implementation(address) function.

token_implementationCall

Function with signature token_implementation() and selector 0x35214d81.

token_implementationReturn

Container type for the return parameters of the token_implementation() function.

Enums

CurveCryptoSwapFactoryCalls

Container for all the CurveCryptoSwapFactory function calls.

CurveCryptoSwapFactoryEvents

Container for all the CurveCryptoSwapFactory events.