Module brontes_classifier::CurveTriCryptoFactory

Generated by the following Solidity interface…

interface CurveTriCryptoFactory {
    event LiquidityGaugeDeployed(address pool, address gauge);
    event TransferOwnership(address _old_owner, address _new_owner);
    event TricryptoPoolDeployed(address pool, string name, string symbol, address weth, address[3] coins, address math, bytes32 salt, uint256 packed_precisions, uint256 packed_A_gamma, uint256 packed_fee_params, uint256 packed_rebalancing_params, uint256 packed_prices, address deployer);
    event UpdateFeeReceiver(address _old_fee_receiver, address _new_fee_receiver);
    event UpdateGaugeImplementation(address _old_gauge_implementation, address _new_gauge_implementation);
    event UpdateMathImplementation(address _old_math_implementation, address _new_math_implementation);
    event UpdatePoolImplementation(uint256 _implemention_id, address _old_pool_implementation, address _new_pool_implementation);
    event UpdateViewsImplementation(address _old_views_implementation, address _new_views_implementation);

    function accept_transfer_ownership() external;
    function admin() external view returns (address);
    function commit_transfer_ownership(address _addr) external;
    function deploy_gauge(address _pool) external returns (address);
    function deploy_pool(string memory _name, string memory _symbol, address[3] memory _coins, address _weth, uint256 implementation_id, uint256 A, uint256 gamma, uint256 mid_fee, uint256 out_fee, uint256 fee_gamma, uint256 allowed_extra_profit, uint256 adjustment_step, uint256 ma_exp_time, uint256[2] memory initial_prices) external returns (address);
    function fee_receiver() external view returns (address);
    function find_pool_for_coins(address _from, address _to) external view returns (address);
    function find_pool_for_coins(address _from, address _to, uint256 i) external view returns (address);
    function future_admin() external view returns (address);
    function gauge_implementation() external view returns (address);
    function get_balances(address _pool) external view returns (uint256[3] memory);
    function get_coin_indices(address _pool, address _from, address _to) external view returns (uint256, uint256);
    function get_coins(address _pool) external view returns (address[3] memory);
    function get_decimals(address _pool) external view returns (uint256[3] memory);
    function get_gauge(address _pool) external view returns (address);
    function get_market_counts(address coin_a, address coin_b) external view returns (uint256);
    function math_implementation() external view returns (address);
    function pool_count() external view returns (uint256);
    function pool_implementations(uint256 arg0) external view returns (address);
    function pool_list(uint256 arg0) external view returns (address);
    function set_fee_receiver(address _fee_receiver) external;
    function set_gauge_implementation(address _gauge_implementation) external;
    function set_math_implementation(address _math_implementation) external;
    function set_pool_implementation(address _pool_implementation, uint256 _implementation_index) external;
    function set_views_implementation(address _views_implementation) external;
    function views_implementation() external view returns (address);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_fee_receiver",
        "type": "address"
      },
      {
        "name": "_admin",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "accept_transfer_ownership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "admin",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "commit_transfer_ownership",
    "inputs": [
      {
        "name": "_addr",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deploy_gauge",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deploy_pool",
    "inputs": [
      {
        "name": "_name",
        "type": "string"
      },
      {
        "name": "_symbol",
        "type": "string"
      },
      {
        "name": "_coins",
        "type": "address[3]"
      },
      {
        "name": "_weth",
        "type": "address"
      },
      {
        "name": "implementation_id",
        "type": "uint256"
      },
      {
        "name": "A",
        "type": "uint256"
      },
      {
        "name": "gamma",
        "type": "uint256"
      },
      {
        "name": "mid_fee",
        "type": "uint256"
      },
      {
        "name": "out_fee",
        "type": "uint256"
      },
      {
        "name": "fee_gamma",
        "type": "uint256"
      },
      {
        "name": "allowed_extra_profit",
        "type": "uint256"
      },
      {
        "name": "adjustment_step",
        "type": "uint256"
      },
      {
        "name": "ma_exp_time",
        "type": "uint256"
      },
      {
        "name": "initial_prices",
        "type": "uint256[2]"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "fee_receiver",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "find_pool_for_coins",
    "inputs": [
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "find_pool_for_coins",
    "inputs": [
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      },
      {
        "name": "i",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_admin",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "gauge_implementation",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_balances",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[3]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_coin_indices",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      },
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      },
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_coins",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address[3]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_decimals",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[3]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_gauge",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_market_counts",
    "inputs": [
      {
        "name": "coin_a",
        "type": "address"
      },
      {
        "name": "coin_b",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "math_implementation",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pool_count",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pool_implementations",
    "inputs": [
      {
        "name": "arg0",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pool_list",
    "inputs": [
      {
        "name": "arg0",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "set_fee_receiver",
    "inputs": [
      {
        "name": "_fee_receiver",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "set_gauge_implementation",
    "inputs": [
      {
        "name": "_gauge_implementation",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "set_math_implementation",
    "inputs": [
      {
        "name": "_math_implementation",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "set_pool_implementation",
    "inputs": [
      {
        "name": "_pool_implementation",
        "type": "address"
      },
      {
        "name": "_implementation_index",
        "type": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "set_views_implementation",
    "inputs": [
      {
        "name": "_views_implementation",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "views_implementation",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "event",
    "name": "LiquidityGaugeDeployed",
    "inputs": [
      {
        "name": "pool",
        "type": "address",
        "indexed": false
      },
      {
        "name": "gauge",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "TransferOwnership",
    "inputs": [
      {
        "name": "_old_owner",
        "type": "address",
        "indexed": false
      },
      {
        "name": "_new_owner",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "TricryptoPoolDeployed",
    "inputs": [
      {
        "name": "pool",
        "type": "address",
        "indexed": false
      },
      {
        "name": "name",
        "type": "string",
        "indexed": false
      },
      {
        "name": "symbol",
        "type": "string",
        "indexed": false
      },
      {
        "name": "weth",
        "type": "address",
        "indexed": false
      },
      {
        "name": "coins",
        "type": "address[3]",
        "indexed": false
      },
      {
        "name": "math",
        "type": "address",
        "indexed": false
      },
      {
        "name": "salt",
        "type": "bytes32",
        "indexed": false
      },
      {
        "name": "packed_precisions",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "packed_A_gamma",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "packed_fee_params",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "packed_rebalancing_params",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "packed_prices",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "deployer",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "UpdateFeeReceiver",
    "inputs": [
      {
        "name": "_old_fee_receiver",
        "type": "address",
        "indexed": false
      },
      {
        "name": "_new_fee_receiver",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "UpdateGaugeImplementation",
    "inputs": [
      {
        "name": "_old_gauge_implementation",
        "type": "address",
        "indexed": false
      },
      {
        "name": "_new_gauge_implementation",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "UpdateMathImplementation",
    "inputs": [
      {
        "name": "_old_math_implementation",
        "type": "address",
        "indexed": false
      },
      {
        "name": "_new_math_implementation",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "UpdatePoolImplementation",
    "inputs": [
      {
        "name": "_implemention_id",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "_old_pool_implementation",
        "type": "address",
        "indexed": false
      },
      {
        "name": "_new_pool_implementation",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "UpdateViewsImplementation",
    "inputs": [
      {
        "name": "_old_views_implementation",
        "type": "address",
        "indexed": false
      },
      {
        "name": "_new_views_implementation",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  }
]

Structs

LiquidityGaugeDeployed

Event with signature LiquidityGaugeDeployed(address,address) and selector 0x656bb34c20491970a8c163f3bd62ead82022b379c3924960ec60f6dbfc5aab3b.

TransferOwnership

Event with signature TransferOwnership(address,address) and selector 0x5c486528ec3e3f0ea91181cff8116f02bfa350e03b8b6f12e00765adbb5af85c.

TricryptoPoolDeployed

Event with signature TricryptoPoolDeployed(address,string,string,address,address[3],address,bytes32,uint256,uint256,uint256,uint256,uint256,address) and selector 0xa307f5d0802489baddec443058a63ce115756de9020e2b07d3e2cd2f21269e2a.

UpdateFeeReceiver

Event with signature UpdateFeeReceiver(address,address) and selector 0x2861448678f0be67f11bfb5481b3e3b4cfeb3acc6126ad60a05f95bfc6530666.

UpdateGaugeImplementation

Event with signature UpdateGaugeImplementation(address,address) and selector 0x1fd705f9c77053962a503f2f2f57f0862b4c3af687c25615c13817a86946c359.

UpdateMathImplementation

Event with signature UpdateMathImplementation(address,address) and selector 0x68fe8fc3ac76ec17e21117df5e854c8c25b7b5f776aad2adc927fdd156bcd6de.

UpdatePoolImplementation

Event with signature UpdatePoolImplementation(uint256,address,address) and selector 0x6a42ef9605e135afaf6ae4f3683b161a3b7369d07c9d52c701ab69553e04c3b6.

UpdateViewsImplementation

Event with signature UpdateViewsImplementation(address,address) and selector 0xd84eb1ea70cda40a6bfaa11f4f69efa10cbc5eb82760b3058f440512ec1d6d1f.

accept_transfer_ownershipCall

Function with signature accept_transfer_ownership() and selector 0xe5ea47b8.

accept_transfer_ownershipReturn

Container type for the return parameters of the accept_transfer_ownership() function.

adminCall

Function with signature admin() and selector 0xf851a440.

adminReturn

Container type for the return parameters of the admin() function.

commit_transfer_ownershipCall

Function with signature commit_transfer_ownership(address) and selector 0x6b441a40.

commit_transfer_ownershipReturn

Container type for the return parameters of the commit_transfer_ownership(address) function.

deploy_gaugeCall

Function with signature deploy_gauge(address) and selector 0x96bebb34.

deploy_gaugeReturn

Container type for the return parameters of the deploy_gauge(address) function.

deploy_poolCall

Function with signature deploy_pool(string,string,address[3],address,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256[2]) and selector 0xaa38b385.

deploy_poolReturn

Container type for the return parameters of the deploy_pool(string,string,address[3],address,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256[2]) function.

fee_receiverCall

Function with signature fee_receiver() and selector 0xcab4d3db.

fee_receiverReturn

Container type for the return parameters of the fee_receiver() function.

find_pool_for_coins_0Call

Function with signature find_pool_for_coins(address,address) and selector 0xa87df06c.

find_pool_for_coins_0Return

Container type for the return parameters of the find_pool_for_coins(address,address) function.

find_pool_for_coins_1Call

Function with signature find_pool_for_coins(address,address,uint256) and selector 0x6982eb0b.

find_pool_for_coins_1Return

Container type for the return parameters of the find_pool_for_coins(address,address,uint256) function.

future_adminCall

Function with signature future_admin() and selector 0x17f7182a.

future_adminReturn

Container type for the return parameters of the future_admin() function.

gauge_implementationCall

Function with signature gauge_implementation() and selector 0x8df24207.

gauge_implementationReturn

Container type for the return parameters of the gauge_implementation() function.

get_balancesCall

Function with signature get_balances(address) and selector 0x92e3cc2d.

get_balancesReturn

Container type for the return parameters of the get_balances(address) function.

get_coin_indicesCall

Function with signature get_coin_indices(address,address,address) and selector 0xeb85226d.

get_coin_indicesReturn

Container type for the return parameters of the get_coin_indices(address,address,address) function.

get_coinsCall

Function with signature get_coins(address) and selector 0x9ac90d3d.

get_coinsReturn

Container type for the return parameters of the get_coins(address) function.

get_decimalsCall

Function with signature get_decimals(address) and selector 0x52b51555.

get_decimalsReturn

Container type for the return parameters of the get_decimals(address) function.

get_gaugeCall

Function with signature get_gauge(address) and selector 0xdaf297b9.

get_gaugeReturn

Container type for the return parameters of the get_gauge(address) function.

get_market_countsCall

Function with signature get_market_counts(address,address) and selector 0xc1856b52.

get_market_countsReturn

Container type for the return parameters of the get_market_counts(address,address) function.

math_implementationCall

Function with signature math_implementation() and selector 0xa13c8f81.

math_implementationReturn

Container type for the return parameters of the math_implementation() function.

pool_countCall

Function with signature pool_count() and selector 0x956aae3a.

pool_countReturn

Container type for the return parameters of the pool_count() function.

pool_implementationsCall

Function with signature pool_implementations(uint256) and selector 0x3273ff47.

pool_implementationsReturn

Container type for the return parameters of the pool_implementations(uint256) function.

pool_listCall

Function with signature pool_list(uint256) and selector 0x3a1d5d8e.

pool_listReturn

Container type for the return parameters of the pool_list(uint256) function.

set_fee_receiverCall

Function with signature set_fee_receiver(address) and selector 0xe41ab771.

set_fee_receiverReturn

Container type for the return parameters of the set_fee_receiver(address) function.

set_gauge_implementationCall

Function with signature set_gauge_implementation(address) and selector 0x8f03182c.

set_gauge_implementationReturn

Container type for the return parameters of the set_gauge_implementation(address) function.

set_math_implementationCall

Function with signature set_math_implementation(address) and selector 0xb07426f4.

set_math_implementationReturn

Container type for the return parameters of the set_math_implementation(address) function.

set_pool_implementationCall

Function with signature set_pool_implementation(address,uint256) and selector 0x6f385ff6.

set_pool_implementationReturn

Container type for the return parameters of the set_pool_implementation(address,uint256) function.

set_views_implementationCall

Function with signature set_views_implementation(address) and selector 0xf6fa937f.

set_views_implementationReturn

Container type for the return parameters of the set_views_implementation(address) function.

views_implementationCall

Function with signature views_implementation() and selector 0xe31593d8.

views_implementationReturn

Container type for the return parameters of the views_implementation() function.

Enums

CurveTriCryptoFactoryCalls

Container for all the CurveTriCryptoFactory function calls.

CurveTriCryptoFactoryEvents

Container for all the CurveTriCryptoFactory events.