Module brontes_classifier::CurveV1MetapoolFactory

Generated by the following Solidity interface…

interface CurveV1MetapoolFactory {
    event BasePoolAdded(address base_pool, address implementat);
    event MetaPoolDeployed(address coin, address base_pool, uint256 A, uint256 fee, address deployer);

    function accept_transfer_ownership() external;
    function add_base_pool(address _base_pool, address _metapool_implementation, address _fee_receiver) external;
    function admin() external view returns (address);
    function base_pool_count() external view returns (uint256);
    function base_pool_list(uint256 arg0) external view returns (address);
    function commit_transfer_ownership(address addr) external;
    function convert_fees() external returns (bool);
    function deploy_metapool(address _base_pool, string memory _name, string memory _symbol, address _coin, uint256 _A, uint256 _fee) external returns (address);
    function fee_receiver(address arg0) external view returns (address);
    function find_pool_for_coins(address _from, address _to) external view returns (address);
    function find_pool_for_coins(address _from, address _to, uint256 i) external view returns (address);
    function future_admin() external view returns (address);
    function get_A(address _pool) external view returns (uint256);
    function get_admin_balances(address _pool) external view returns (uint256[2] memory);
    function get_balances(address _pool) external view returns (uint256[2] memory);
    function get_coin_indices(address _pool, address _from, address _to) external view returns (int128, int128, bool);
    function get_coins(address _pool) external view returns (address[2] memory);
    function get_decimals(address _pool) external view returns (uint256[2] memory);
    function get_fees(address _pool) external view returns (uint256, uint256);
    function get_n_coins(address _pool) external view returns (uint256, uint256);
    function get_rates(address _pool) external view returns (uint256[2] memory);
    function get_underlying_balances(address _pool) external view returns (uint256[8] memory);
    function get_underlying_coins(address _pool) external view returns (address[8] memory);
    function get_underlying_decimals(address _pool) external view returns (uint256[8] memory);
    function pool_count() external view returns (uint256);
    function pool_list(uint256 arg0) external view returns (address);
    function set_fee_receiver(address _base_pool, address _fee_receiver) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "accept_transfer_ownership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "add_base_pool",
    "inputs": [
      {
        "name": "_base_pool",
        "type": "address"
      },
      {
        "name": "_metapool_implementation",
        "type": "address"
      },
      {
        "name": "_fee_receiver",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "admin",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "base_pool_count",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "base_pool_list",
    "inputs": [
      {
        "name": "arg0",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "commit_transfer_ownership",
    "inputs": [
      {
        "name": "addr",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "convert_fees",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deploy_metapool",
    "inputs": [
      {
        "name": "_base_pool",
        "type": "address"
      },
      {
        "name": "_name",
        "type": "string"
      },
      {
        "name": "_symbol",
        "type": "string"
      },
      {
        "name": "_coin",
        "type": "address"
      },
      {
        "name": "_A",
        "type": "uint256"
      },
      {
        "name": "_fee",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "fee_receiver",
    "inputs": [
      {
        "name": "arg0",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "find_pool_for_coins",
    "inputs": [
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "find_pool_for_coins",
    "inputs": [
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      },
      {
        "name": "i",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_admin",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_A",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_admin_balances",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[2]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_balances",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[2]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_coin_indices",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      },
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "int128"
      },
      {
        "name": "",
        "type": "int128"
      },
      {
        "name": "",
        "type": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_coins",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address[2]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_decimals",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[2]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_fees",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      },
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_n_coins",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      },
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_rates",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[2]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_underlying_balances",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[8]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_underlying_coins",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address[8]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_underlying_decimals",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[8]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pool_count",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pool_list",
    "inputs": [
      {
        "name": "arg0",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "set_fee_receiver",
    "inputs": [
      {
        "name": "_base_pool",
        "type": "address"
      },
      {
        "name": "_fee_receiver",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "BasePoolAdded",
    "inputs": [
      {
        "name": "base_pool",
        "type": "address",
        "indexed": false
      },
      {
        "name": "implementat",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "MetaPoolDeployed",
    "inputs": [
      {
        "name": "coin",
        "type": "address",
        "indexed": false
      },
      {
        "name": "base_pool",
        "type": "address",
        "indexed": false
      },
      {
        "name": "A",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "deployer",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  }
]

Structs

BasePoolAdded

Event with signature BasePoolAdded(address,address) and selector 0xd1951e19788b24c3c75b3c6c19a795cd312b1523d09076fc41e432cd3ccdc14c.

MetaPoolDeployed

Event with signature MetaPoolDeployed(address,address,uint256,uint256,address) and selector 0x01f31cd2abdeb4e5e10ba500f2db0f937d9e8c735ab04681925441b4ea37eda5.

accept_transfer_ownershipCall

Function with signature accept_transfer_ownership() and selector 0xe5ea47b8.

accept_transfer_ownershipReturn

Container type for the return parameters of the accept_transfer_ownership() function.

add_base_poolCall

Function with signature add_base_pool(address,address,address) and selector 0x640ada50.

add_base_poolReturn

Container type for the return parameters of the add_base_pool(address,address,address) function.

adminCall

Function with signature admin() and selector 0xf851a440.

adminReturn

Container type for the return parameters of the admin() function.

base_pool_countCall

Function with signature base_pool_count() and selector 0xde5e4a3b.

base_pool_countReturn

Container type for the return parameters of the base_pool_count() function.

base_pool_listCall

Function with signature base_pool_list(uint256) and selector 0x22fe5671.

base_pool_listReturn

Container type for the return parameters of the base_pool_list(uint256) function.

commit_transfer_ownershipCall

Function with signature commit_transfer_ownership(address) and selector 0x6b441a40.

commit_transfer_ownershipReturn

Container type for the return parameters of the commit_transfer_ownership(address) function.

convert_feesCall

Function with signature convert_fees() and selector 0xdc125577.

convert_feesReturn

Container type for the return parameters of the convert_fees() function.

deploy_metapoolCall

Function with signature deploy_metapool(address,string,string,address,uint256,uint256) and selector 0xe339eb4f.

deploy_metapoolReturn

Container type for the return parameters of the deploy_metapool(address,string,string,address,uint256,uint256) function.

fee_receiverCall

Function with signature fee_receiver(address) and selector 0x72b38d98.

fee_receiverReturn

Container type for the return parameters of the fee_receiver(address) function.

find_pool_for_coins_0Call

Function with signature find_pool_for_coins(address,address) and selector 0xa87df06c.

find_pool_for_coins_0Return

Container type for the return parameters of the find_pool_for_coins(address,address) function.

find_pool_for_coins_1Call

Function with signature find_pool_for_coins(address,address,uint256) and selector 0x6982eb0b.

find_pool_for_coins_1Return

Container type for the return parameters of the find_pool_for_coins(address,address,uint256) function.

future_adminCall

Function with signature future_admin() and selector 0x17f7182a.

future_adminReturn

Container type for the return parameters of the future_admin() function.

get_ACall

Function with signature get_A(address) and selector 0x55b30b19.

get_AReturn

Container type for the return parameters of the get_A(address) function.

get_admin_balancesCall

Function with signature get_admin_balances(address) and selector 0xc11e45b8.

get_admin_balancesReturn

Container type for the return parameters of the get_admin_balances(address) function.

get_balancesCall

Function with signature get_balances(address) and selector 0x92e3cc2d.

get_balancesReturn

Container type for the return parameters of the get_balances(address) function.

get_coin_indicesCall

Function with signature get_coin_indices(address,address,address) and selector 0xeb85226d.

get_coin_indicesReturn

Container type for the return parameters of the get_coin_indices(address,address,address) function.

get_coinsCall

Function with signature get_coins(address) and selector 0x9ac90d3d.

get_coinsReturn

Container type for the return parameters of the get_coins(address) function.

get_decimalsCall

Function with signature get_decimals(address) and selector 0x52b51555.

get_decimalsReturn

Container type for the return parameters of the get_decimals(address) function.

get_feesCall

Function with signature get_fees(address) and selector 0x7cdb72b0.

get_feesReturn

Container type for the return parameters of the get_fees(address) function.

get_n_coinsCall

Function with signature get_n_coins(address) and selector 0x940494f1.

get_n_coinsReturn

Container type for the return parameters of the get_n_coins(address) function.

get_ratesCall

Function with signature get_rates(address) and selector 0xce99e45a.

get_ratesReturn

Container type for the return parameters of the get_rates(address) function.

get_underlying_balancesCall

Function with signature get_underlying_balances(address) and selector 0x59f4f351.

get_underlying_balancesReturn

Container type for the return parameters of the get_underlying_balances(address) function.

get_underlying_coinsCall

Function with signature get_underlying_coins(address) and selector 0xa77576ef.

get_underlying_coinsReturn

Container type for the return parameters of the get_underlying_coins(address) function.

get_underlying_decimalsCall

Function with signature get_underlying_decimals(address) and selector 0x4cb088f1.

get_underlying_decimalsReturn

Container type for the return parameters of the get_underlying_decimals(address) function.

pool_countCall

Function with signature pool_count() and selector 0x956aae3a.

pool_countReturn

Container type for the return parameters of the pool_count() function.

pool_listCall

Function with signature pool_list(uint256) and selector 0x3a1d5d8e.

pool_listReturn

Container type for the return parameters of the pool_list(uint256) function.

set_fee_receiverCall

Function with signature set_fee_receiver(address,address) and selector 0x36d2b77a.

set_fee_receiverReturn

Container type for the return parameters of the set_fee_receiver(address,address) function.

Enums

CurveV1MetapoolFactoryCalls

Container for all the CurveV1MetapoolFactory function calls.

CurveV1MetapoolFactoryEvents

Container for all the CurveV1MetapoolFactory events.