Module brontes_classifier::CurvecrvUSDFactory

Generated by the following Solidity interface…

interface CurvecrvUSDFactory {
    event BasePoolAdded(address base_pool);
    event LiquidityGaugeDeployed(address pool, address gauge);
    event MetaPoolDeployed(address coin, address base_pool, uint256 A, uint256 fee, address deployer);
    event PlainPoolDeployed(address[4] coins, uint256 A, uint256 fee, address deployer, address pool);

    function accept_transfer_ownership() external;
    function add_base_pool(address _base_pool, address _fee_receiver, uint256 _asset_type, address[10] memory _implementations) external;
    function add_existing_metapools(address[10] memory _pools) external returns (bool);
    function add_token_to_whitelist(address coin, bool _add) external;
    function admin() external view returns (address);
    function base_pool_assets(address arg0) external view returns (bool);
    function base_pool_count() external view returns (uint256);
    function base_pool_list(uint256 arg0) external view returns (address);
    function batch_set_pool_asset_type(address[32] memory _pools, uint256[32] memory _asset_types) external;
    function commit_transfer_ownership(address _addr) external;
    function convert_metapool_fees() external returns (bool);
    function deploy_gauge(address _pool) external returns (address);
    function deploy_metapool(address _base_pool, string memory _name, string memory _symbol, address _coin, uint256 _A, uint256 _fee) external returns (address);
    function deploy_metapool(address _base_pool, string memory _name, string memory _symbol, address _coin, uint256 _A, uint256 _fee, uint256 _implementation_idx) external returns (address);
    function deploy_plain_pool(string memory _name, string memory _symbol, address[4] memory _coins, uint256 _A, uint256 _fee) external returns (address);
    function deploy_plain_pool(string memory _name, string memory _symbol, address[4] memory _coins, uint256 _A, uint256 _fee, uint256 _asset_type) external returns (address);
    function deploy_plain_pool(string memory _name, string memory _symbol, address[4] memory _coins, uint256 _A, uint256 _fee, uint256 _asset_type, uint256 _implementation_idx) external returns (address);
    function find_pool_for_coins(address _from, address _to) external view returns (address);
    function find_pool_for_coins(address _from, address _to, uint256 i) external view returns (address);
    function future_admin() external view returns (address);
    function gauge_implementation() external view returns (address);
    function get_A(address _pool) external view returns (uint256);
    function get_admin_balances(address _pool) external view returns (uint256[4] memory);
    function get_balances(address _pool) external view returns (uint256[4] memory);
    function get_base_pool(address _pool) external view returns (address);
    function get_coin_indices(address _pool, address _from, address _to) external view returns (int128, int128, bool);
    function get_coins(address _pool) external view returns (address[4] memory);
    function get_decimals(address _pool) external view returns (uint256[4] memory);
    function get_fee_receiver(address _pool) external view returns (address);
    function get_fees(address _pool) external view returns (uint256, uint256);
    function get_gauge(address _pool) external view returns (address);
    function get_implementation_address(address _pool) external view returns (address);
    function get_meta_n_coins(address _pool) external view returns (uint256, uint256);
    function get_metapool_rates(address _pool) external view returns (uint256[2] memory);
    function get_n_coins(address _pool) external view returns (uint256);
    function get_pool_asset_type(address _pool) external view returns (uint256);
    function get_underlying_balances(address _pool) external view returns (uint256[8] memory);
    function get_underlying_coins(address _pool) external view returns (address[8] memory);
    function get_underlying_decimals(address _pool) external view returns (uint256[8] memory);
    function is_meta(address _pool) external view returns (bool);
    function manager() external view returns (address);
    function metapool_implementations(address _base_pool) external view returns (address[10] memory);
    function plain_implementations(uint256 arg0, uint256 arg1) external view returns (address);
    function plain_whitelist(address arg0) external view returns (bool);
    function pool_count() external view returns (uint256);
    function pool_list(uint256 arg0) external view returns (address);
    function set_fee_receiver(address _base_pool, address _fee_receiver) external;
    function set_gauge_implementation(address _gauge_implementation) external;
    function set_manager(address _manager) external;
    function set_metapool_implementations(address _base_pool, address[10] memory _implementations) external;
    function set_plain_implementations(uint256 _n_coins, address[10] memory _implementations) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_fee_receiver",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "accept_transfer_ownership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "add_base_pool",
    "inputs": [
      {
        "name": "_base_pool",
        "type": "address"
      },
      {
        "name": "_fee_receiver",
        "type": "address"
      },
      {
        "name": "_asset_type",
        "type": "uint256"
      },
      {
        "name": "_implementations",
        "type": "address[10]"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "add_existing_metapools",
    "inputs": [
      {
        "name": "_pools",
        "type": "address[10]"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "add_token_to_whitelist",
    "inputs": [
      {
        "name": "coin",
        "type": "address"
      },
      {
        "name": "_add",
        "type": "bool"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "admin",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "base_pool_assets",
    "inputs": [
      {
        "name": "arg0",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "base_pool_count",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "base_pool_list",
    "inputs": [
      {
        "name": "arg0",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "batch_set_pool_asset_type",
    "inputs": [
      {
        "name": "_pools",
        "type": "address[32]"
      },
      {
        "name": "_asset_types",
        "type": "uint256[32]"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "commit_transfer_ownership",
    "inputs": [
      {
        "name": "_addr",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "convert_metapool_fees",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deploy_gauge",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deploy_metapool",
    "inputs": [
      {
        "name": "_base_pool",
        "type": "address"
      },
      {
        "name": "_name",
        "type": "string"
      },
      {
        "name": "_symbol",
        "type": "string"
      },
      {
        "name": "_coin",
        "type": "address"
      },
      {
        "name": "_A",
        "type": "uint256"
      },
      {
        "name": "_fee",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deploy_metapool",
    "inputs": [
      {
        "name": "_base_pool",
        "type": "address"
      },
      {
        "name": "_name",
        "type": "string"
      },
      {
        "name": "_symbol",
        "type": "string"
      },
      {
        "name": "_coin",
        "type": "address"
      },
      {
        "name": "_A",
        "type": "uint256"
      },
      {
        "name": "_fee",
        "type": "uint256"
      },
      {
        "name": "_implementation_idx",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deploy_plain_pool",
    "inputs": [
      {
        "name": "_name",
        "type": "string"
      },
      {
        "name": "_symbol",
        "type": "string"
      },
      {
        "name": "_coins",
        "type": "address[4]"
      },
      {
        "name": "_A",
        "type": "uint256"
      },
      {
        "name": "_fee",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deploy_plain_pool",
    "inputs": [
      {
        "name": "_name",
        "type": "string"
      },
      {
        "name": "_symbol",
        "type": "string"
      },
      {
        "name": "_coins",
        "type": "address[4]"
      },
      {
        "name": "_A",
        "type": "uint256"
      },
      {
        "name": "_fee",
        "type": "uint256"
      },
      {
        "name": "_asset_type",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "deploy_plain_pool",
    "inputs": [
      {
        "name": "_name",
        "type": "string"
      },
      {
        "name": "_symbol",
        "type": "string"
      },
      {
        "name": "_coins",
        "type": "address[4]"
      },
      {
        "name": "_A",
        "type": "uint256"
      },
      {
        "name": "_fee",
        "type": "uint256"
      },
      {
        "name": "_asset_type",
        "type": "uint256"
      },
      {
        "name": "_implementation_idx",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "find_pool_for_coins",
    "inputs": [
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "find_pool_for_coins",
    "inputs": [
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      },
      {
        "name": "i",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "future_admin",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "gauge_implementation",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_A",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_admin_balances",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[4]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_balances",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[4]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_base_pool",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_coin_indices",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      },
      {
        "name": "_from",
        "type": "address"
      },
      {
        "name": "_to",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "int128"
      },
      {
        "name": "",
        "type": "int128"
      },
      {
        "name": "",
        "type": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_coins",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address[4]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_decimals",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[4]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_fee_receiver",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_fees",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      },
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_gauge",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_implementation_address",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_meta_n_coins",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      },
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_metapool_rates",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[2]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_n_coins",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_pool_asset_type",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_underlying_balances",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[8]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_underlying_coins",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address[8]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "get_underlying_decimals",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256[8]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "is_meta",
    "inputs": [
      {
        "name": "_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "manager",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "metapool_implementations",
    "inputs": [
      {
        "name": "_base_pool",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address[10]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "plain_implementations",
    "inputs": [
      {
        "name": "arg0",
        "type": "uint256"
      },
      {
        "name": "arg1",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "plain_whitelist",
    "inputs": [
      {
        "name": "arg0",
        "type": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pool_count",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pool_list",
    "inputs": [
      {
        "name": "arg0",
        "type": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "set_fee_receiver",
    "inputs": [
      {
        "name": "_base_pool",
        "type": "address"
      },
      {
        "name": "_fee_receiver",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "set_gauge_implementation",
    "inputs": [
      {
        "name": "_gauge_implementation",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "set_manager",
    "inputs": [
      {
        "name": "_manager",
        "type": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "set_metapool_implementations",
    "inputs": [
      {
        "name": "_base_pool",
        "type": "address"
      },
      {
        "name": "_implementations",
        "type": "address[10]"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "set_plain_implementations",
    "inputs": [
      {
        "name": "_n_coins",
        "type": "uint256"
      },
      {
        "name": "_implementations",
        "type": "address[10]"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "BasePoolAdded",
    "inputs": [
      {
        "name": "base_pool",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "LiquidityGaugeDeployed",
    "inputs": [
      {
        "name": "pool",
        "type": "address",
        "indexed": false
      },
      {
        "name": "gauge",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "MetaPoolDeployed",
    "inputs": [
      {
        "name": "coin",
        "type": "address",
        "indexed": false
      },
      {
        "name": "base_pool",
        "type": "address",
        "indexed": false
      },
      {
        "name": "A",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "deployer",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "PlainPoolDeployed",
    "inputs": [
      {
        "name": "coins",
        "type": "address[4]",
        "indexed": false
      },
      {
        "name": "A",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "fee",
        "type": "uint256",
        "indexed": false
      },
      {
        "name": "deployer",
        "type": "address",
        "indexed": false
      },
      {
        "name": "pool",
        "type": "address",
        "indexed": false
      }
    ],
    "anonymous": false
  }
]

Structs

BasePoolAdded

Event with signature BasePoolAdded(address) and selector 0xcc6afdfec79da6be08142ecee25cf14b665961e25d30d8eba45959be9547635f.

LiquidityGaugeDeployed

Event with signature LiquidityGaugeDeployed(address,address) and selector 0x656bb34c20491970a8c163f3bd62ead82022b379c3924960ec60f6dbfc5aab3b.

MetaPoolDeployed

Event with signature MetaPoolDeployed(address,address,uint256,uint256,address) and selector 0x01f31cd2abdeb4e5e10ba500f2db0f937d9e8c735ab04681925441b4ea37eda5.

PlainPoolDeployed

Event with signature PlainPoolDeployed(address[4],uint256,uint256,address,address) and selector 0xb8f6972d6e56d21c47621efd7f02fe68f07a17c999c42245b3abd300f34d61eb.

accept_transfer_ownershipCall

Function with signature accept_transfer_ownership() and selector 0xe5ea47b8.

accept_transfer_ownershipReturn

Container type for the return parameters of the accept_transfer_ownership() function.

add_base_poolCall

Function with signature add_base_pool(address,address,uint256,address[10]) and selector 0x2fc05653.

add_base_poolReturn

Container type for the return parameters of the add_base_pool(address,address,uint256,address[10]) function.

add_existing_metapoolsCall

Function with signature add_existing_metapools(address[10]) and selector 0xdb89fabc.

add_existing_metapoolsReturn

Container type for the return parameters of the add_existing_metapools(address[10]) function.

add_token_to_whitelistCall

Function with signature add_token_to_whitelist(address,bool) and selector 0x2a91f816.

add_token_to_whitelistReturn

Container type for the return parameters of the add_token_to_whitelist(address,bool) function.

adminCall

Function with signature admin() and selector 0xf851a440.

adminReturn

Container type for the return parameters of the admin() function.

base_pool_assetsCall

Function with signature base_pool_assets(address) and selector 0x10a002df.

base_pool_assetsReturn

Container type for the return parameters of the base_pool_assets(address) function.

base_pool_countCall

Function with signature base_pool_count() and selector 0xde5e4a3b.

base_pool_countReturn

Container type for the return parameters of the base_pool_count() function.

base_pool_listCall

Function with signature base_pool_list(uint256) and selector 0x22fe5671.

base_pool_listReturn

Container type for the return parameters of the base_pool_list(uint256) function.

batch_set_pool_asset_typeCall

Function with signature batch_set_pool_asset_type(address[32],uint256[32]) and selector 0x7542f078.

batch_set_pool_asset_typeReturn

Container type for the return parameters of the batch_set_pool_asset_type(address[32],uint256[32]) function.

commit_transfer_ownershipCall

Function with signature commit_transfer_ownership(address) and selector 0x6b441a40.

commit_transfer_ownershipReturn

Container type for the return parameters of the commit_transfer_ownership(address) function.

convert_metapool_feesCall

Function with signature convert_metapool_fees() and selector 0xbcc981d2.

convert_metapool_feesReturn

Container type for the return parameters of the convert_metapool_fees() function.

deploy_gaugeCall

Function with signature deploy_gauge(address) and selector 0x96bebb34.

deploy_gaugeReturn

Container type for the return parameters of the deploy_gauge(address) function.

deploy_metapool_0Call

Function with signature deploy_metapool(address,string,string,address,uint256,uint256) and selector 0xe339eb4f.

deploy_metapool_0Return

Container type for the return parameters of the deploy_metapool(address,string,string,address,uint256,uint256) function.

deploy_metapool_1Call

Function with signature deploy_metapool(address,string,string,address,uint256,uint256,uint256) and selector 0xde7fe3bf.

deploy_metapool_1Return

Container type for the return parameters of the deploy_metapool(address,string,string,address,uint256,uint256,uint256) function.

deploy_plain_pool_0Call

Function with signature deploy_plain_pool(string,string,address[4],uint256,uint256) and selector 0xcd419bb5.

deploy_plain_pool_0Return

Container type for the return parameters of the deploy_plain_pool(string,string,address[4],uint256,uint256) function.

deploy_plain_pool_1Call

Function with signature deploy_plain_pool(string,string,address[4],uint256,uint256,uint256) and selector 0x5c16487b.

deploy_plain_pool_1Return

Container type for the return parameters of the deploy_plain_pool(string,string,address[4],uint256,uint256,uint256) function.

deploy_plain_pool_2Call

Function with signature deploy_plain_pool(string,string,address[4],uint256,uint256,uint256,uint256) and selector 0x52f2db69.

deploy_plain_pool_2Return

Container type for the return parameters of the deploy_plain_pool(string,string,address[4],uint256,uint256,uint256,uint256) function.

find_pool_for_coins_0Call

Function with signature find_pool_for_coins(address,address) and selector 0xa87df06c.

find_pool_for_coins_0Return

Container type for the return parameters of the find_pool_for_coins(address,address) function.

find_pool_for_coins_1Call

Function with signature find_pool_for_coins(address,address,uint256) and selector 0x6982eb0b.

find_pool_for_coins_1Return

Container type for the return parameters of the find_pool_for_coins(address,address,uint256) function.

future_adminCall

Function with signature future_admin() and selector 0x17f7182a.

future_adminReturn

Container type for the return parameters of the future_admin() function.

gauge_implementationCall

Function with signature gauge_implementation() and selector 0x8df24207.

gauge_implementationReturn

Container type for the return parameters of the gauge_implementation() function.

get_ACall

Function with signature get_A(address) and selector 0x55b30b19.

get_AReturn

Container type for the return parameters of the get_A(address) function.

get_admin_balancesCall

Function with signature get_admin_balances(address) and selector 0xc11e45b8.

get_admin_balancesReturn

Container type for the return parameters of the get_admin_balances(address) function.

get_balancesCall

Function with signature get_balances(address) and selector 0x92e3cc2d.

get_balancesReturn

Container type for the return parameters of the get_balances(address) function.

get_base_poolCall

Function with signature get_base_pool(address) and selector 0x6f20d6dd.

get_base_poolReturn

Container type for the return parameters of the get_base_pool(address) function.

get_coin_indicesCall

Function with signature get_coin_indices(address,address,address) and selector 0xeb85226d.

get_coin_indicesReturn

Container type for the return parameters of the get_coin_indices(address,address,address) function.

get_coinsCall

Function with signature get_coins(address) and selector 0x9ac90d3d.

get_coinsReturn

Container type for the return parameters of the get_coins(address) function.

get_decimalsCall

Function with signature get_decimals(address) and selector 0x52b51555.

get_decimalsReturn

Container type for the return parameters of the get_decimals(address) function.

get_fee_receiverCall

Function with signature get_fee_receiver(address) and selector 0x154aa8f5.

get_fee_receiverReturn

Container type for the return parameters of the get_fee_receiver(address) function.

get_feesCall

Function with signature get_fees(address) and selector 0x7cdb72b0.

get_feesReturn

Container type for the return parameters of the get_fees(address) function.

get_gaugeCall

Function with signature get_gauge(address) and selector 0xdaf297b9.

get_gaugeReturn

Container type for the return parameters of the get_gauge(address) function.

get_implementation_addressCall

Function with signature get_implementation_address(address) and selector 0x510d98a4.

get_implementation_addressReturn

Container type for the return parameters of the get_implementation_address(address) function.

get_meta_n_coinsCall

Function with signature get_meta_n_coins(address) and selector 0xeb73f37d.

get_meta_n_coinsReturn

Container type for the return parameters of the get_meta_n_coins(address) function.

get_metapool_ratesCall

Function with signature get_metapool_rates(address) and selector 0x06d8f160.

get_metapool_ratesReturn

Container type for the return parameters of the get_metapool_rates(address) function.

get_n_coinsCall

Function with signature get_n_coins(address) and selector 0x940494f1.

get_n_coinsReturn

Container type for the return parameters of the get_n_coins(address) function.

get_pool_asset_typeCall

Function with signature get_pool_asset_type(address) and selector 0x66d3966c.

get_pool_asset_typeReturn

Container type for the return parameters of the get_pool_asset_type(address) function.

get_underlying_balancesCall

Function with signature get_underlying_balances(address) and selector 0x59f4f351.

get_underlying_balancesReturn

Container type for the return parameters of the get_underlying_balances(address) function.

get_underlying_coinsCall

Function with signature get_underlying_coins(address) and selector 0xa77576ef.

get_underlying_coinsReturn

Container type for the return parameters of the get_underlying_coins(address) function.

get_underlying_decimalsCall

Function with signature get_underlying_decimals(address) and selector 0x4cb088f1.

get_underlying_decimalsReturn

Container type for the return parameters of the get_underlying_decimals(address) function.

is_metaCall

Function with signature is_meta(address) and selector 0xe4d332a9.

is_metaReturn

Container type for the return parameters of the is_meta(address) function.

managerCall

Function with signature manager() and selector 0x481c6a75.

managerReturn

Container type for the return parameters of the manager() function.

metapool_implementationsCall

Function with signature metapool_implementations(address) and selector 0x970fa3f3.

metapool_implementationsReturn

Container type for the return parameters of the metapool_implementations(address) function.

plain_implementationsCall

Function with signature plain_implementations(uint256,uint256) and selector 0x31a4f865.

plain_implementationsReturn

Container type for the return parameters of the plain_implementations(uint256,uint256) function.

plain_whitelistCall

Function with signature plain_whitelist(address) and selector 0xde7af82e.

plain_whitelistReturn

Container type for the return parameters of the plain_whitelist(address) function.

pool_countCall

Function with signature pool_count() and selector 0x956aae3a.

pool_countReturn

Container type for the return parameters of the pool_count() function.

pool_listCall

Function with signature pool_list(uint256) and selector 0x3a1d5d8e.

pool_listReturn

Container type for the return parameters of the pool_list(uint256) function.

set_fee_receiverCall

Function with signature set_fee_receiver(address,address) and selector 0x36d2b77a.

set_fee_receiverReturn

Container type for the return parameters of the set_fee_receiver(address,address) function.

set_gauge_implementationCall

Function with signature set_gauge_implementation(address) and selector 0x8f03182c.

set_gauge_implementationReturn

Container type for the return parameters of the set_gauge_implementation(address) function.

set_managerCall

Function with signature set_manager(address) and selector 0x9aece83e.

set_managerReturn

Container type for the return parameters of the set_manager(address) function.

set_metapool_implementationsCall

Function with signature set_metapool_implementations(address,address[10]) and selector 0xcb956b46.

set_metapool_implementationsReturn

Container type for the return parameters of the set_metapool_implementations(address,address[10]) function.

set_plain_implementationsCall

Function with signature set_plain_implementations(uint256,address[10]) and selector 0x9ddbf4b9.

set_plain_implementationsReturn

Container type for the return parameters of the set_plain_implementations(uint256,address[10]) function.

Enums

CurvecrvUSDFactoryCalls

Container for all the CurvecrvUSDFactory function calls.

CurvecrvUSDFactoryEvents

Container for all the CurvecrvUSDFactory events.