Module brontes_classifier::MakerPSM

Generated by the following Solidity interface…

interface MakerPSM {
    event BuyGem(address indexed owner, uint256 value, uint256 fee);
    event Deny(address user);
    event File(bytes32 indexed what, uint256 data);
    event Rely(address user);
    event SellGem(address indexed owner, uint256 value, uint256 fee);

    constructor(address gemJoin_, address daiJoin_, address vow_);

    function buyGem(address usr, uint256 gemAmt) external;
    function dai() external view returns (address);
    function daiJoin() external view returns (address);
    function deny(address usr) external;
    function file(bytes32 what, uint256 data) external;
    function gemJoin() external view returns (address);
    function hope(address usr) external;
    function ilk() external view returns (bytes32);
    function nope(address usr) external;
    function rely(address usr) external;
    function sellGem(address usr, uint256 gemAmt) external;
    function tin() external view returns (uint256);
    function tout() external view returns (uint256);
    function vat() external view returns (address);
    function vow() external view returns (address);
    function wards(address) external view returns (uint256);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "gemJoin_",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "daiJoin_",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "vow_",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "buyGem",
    "inputs": [
      {
        "name": "usr",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "gemAmt",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "dai",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract DaiAbstract"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "daiJoin",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract DaiJoinAbstract"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "deny",
    "inputs": [
      {
        "name": "usr",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "file",
    "inputs": [
      {
        "name": "what",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "data",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "gemJoin",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract AuthGemJoinAbstract"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "hope",
    "inputs": [
      {
        "name": "usr",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "ilk",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "nope",
    "inputs": [
      {
        "name": "usr",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "rely",
    "inputs": [
      {
        "name": "usr",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "sellGem",
    "inputs": [
      {
        "name": "usr",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "gemAmt",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "tin",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "tout",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "vat",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract VatAbstract"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "vow",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "wards",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "event",
    "name": "BuyGem",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "value",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "fee",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Deny",
    "inputs": [
      {
        "name": "user",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "File",
    "inputs": [
      {
        "name": "what",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "data",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Rely",
    "inputs": [
      {
        "name": "user",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "SellGem",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "value",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "fee",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  }
]

Structs

BuyGem

Event with signature BuyGem(address,uint256,uint256) and selector 0x085d06ecf4c34b237767a31c0888e121d89546a77f186f1987c6b8715e1a8caa.

Deny

Event with signature Deny(address) and selector 0x184450df2e323acec0ed3b5c7531b81f9b4cdef7914dfd4c0a4317416bb5251b.

File

Event with signature File(bytes32,uint256) and selector 0xe986e40cc8c151830d4f61050f4fb2e4add8567caad2d5f5496f9158e91fe4c7.

Rely

Event with signature Rely(address) and selector 0xdd0e34038ac38b2a1ce960229778ac48a8719bc900b6c4f8d0475c6e8b385a60.

SellGem

Event with signature SellGem(address,uint256,uint256) and selector 0xef75f5a47cc9a929968796ceb84f19e7541617b4577f2c228ea95200e1572081.

buyGemCall

Function with signature buyGem(address,uint256) and selector 0x8d7ef9bb.

buyGemReturn

Container type for the return parameters of the buyGem(address,uint256) function.

constructorCall

Constructor`.

daiCall

Function with signature dai() and selector 0xf4b9fa75.

daiJoinCall

Function with signature daiJoin() and selector 0xc11645bc.

daiJoinReturn

Container type for the return parameters of the daiJoin() function.

daiReturn

Container type for the return parameters of the dai() function.

denyCall

Function with signature deny(address) and selector 0x9c52a7f1.

denyReturn

Container type for the return parameters of the deny(address) function.

fileCall

Function with signature file(bytes32,uint256) and selector 0x29ae8114.

fileReturn

Container type for the return parameters of the file(bytes32,uint256) function.

gemJoinCall

Function with signature gemJoin() and selector 0x01664f66.

gemJoinReturn

Container type for the return parameters of the gemJoin() function.

hopeCall

Function with signature hope(address) and selector 0xa3b22fc4.

hopeReturn

Container type for the return parameters of the hope(address) function.

ilkCall

Function with signature ilk() and selector 0xc5ce281e.

ilkReturn

Container type for the return parameters of the ilk() function.

nopeCall

Function with signature nope(address) and selector 0xdc4d20fa.

nopeReturn

Container type for the return parameters of the nope(address) function.

relyCall

Function with signature rely(address) and selector 0x65fae35e.

relyReturn

Container type for the return parameters of the rely(address) function.

sellGemCall

Function with signature sellGem(address,uint256) and selector 0x95991276.

sellGemReturn

Container type for the return parameters of the sellGem(address,uint256) function.

tinCall

Function with signature tin() and selector 0x568d4b6f.

tinReturn

Container type for the return parameters of the tin() function.

toutCall

Function with signature tout() and selector 0xfae036d5.

toutReturn

Container type for the return parameters of the tout() function.

vatCall

Function with signature vat() and selector 0x36569e77.

vatReturn

Container type for the return parameters of the vat() function.

vowCall

Function with signature vow() and selector 0x626cb3c5.

vowReturn

Container type for the return parameters of the vow() function.

wardsCall

Function with signature wards(address) and selector 0xbf353dbb.

wardsReturn

Container type for the return parameters of the wards(address) function.

Enums

MakerPSMCalls

Container for all the MakerPSM function calls.

MakerPSMEvents

Container for all the MakerPSM events.