Module brontes_classifier::PancakeSwapV3

Generated by the following Solidity interface…

interface PancakeSwapV3 {
    event Burn(address indexed owner, int24 indexed tickLower, int24 indexed tickUpper, uint128 amount, uint256 amount0, uint256 amount1);
    event Collect(address indexed owner, address recipient, int24 indexed tickLower, int24 indexed tickUpper, uint128 amount0, uint128 amount1);
    event CollectProtocol(address indexed sender, address indexed recipient, uint128 amount0, uint128 amount1);
    event Flash(address indexed sender, address indexed recipient, uint256 amount0, uint256 amount1, uint256 paid0, uint256 paid1);
    event IncreaseObservationCardinalityNext(uint16 observationCardinalityNextOld, uint16 observationCardinalityNextNew);
    event Initialize(uint160 sqrtPriceX96, int24 tick);
    event Mint(address sender, address indexed owner, int24 indexed tickLower, int24 indexed tickUpper, uint128 amount, uint256 amount0, uint256 amount1);
    event SetFeeProtocol(uint32 feeProtocol0Old, uint32 feeProtocol1Old, uint32 feeProtocol0New, uint32 feeProtocol1New);
    event SetLmPoolEvent(address addr);
    event Swap(address indexed sender, address indexed recipient, int256 amount0, int256 amount1, uint160 sqrtPriceX96, uint128 liquidity, int24 tick, uint128 protocolFeesToken0, uint128 protocolFeesToken1);

    constructor();

    function burn(int24 tickLower, int24 tickUpper, uint128 amount) external returns (uint256 amount0, uint256 amount1);
    function collect(address recipient, int24 tickLower, int24 tickUpper, uint128 amount0Requested, uint128 amount1Requested) external returns (uint128 amount0, uint128 amount1);
    function collectProtocol(address recipient, uint128 amount0Requested, uint128 amount1Requested) external returns (uint128 amount0, uint128 amount1);
    function factory() external view returns (address);
    function fee() external view returns (uint24);
    function feeGrowthGlobal0X128() external view returns (uint256);
    function feeGrowthGlobal1X128() external view returns (uint256);
    function flash(address recipient, uint256 amount0, uint256 amount1, bytes memory data) external;
    function increaseObservationCardinalityNext(uint16 observationCardinalityNext) external;
    function initialize(uint160 sqrtPriceX96) external;
    function liquidity() external view returns (uint128);
    function lmPool() external view returns (address);
    function maxLiquidityPerTick() external view returns (uint128);
    function mint(address recipient, int24 tickLower, int24 tickUpper, uint128 amount, bytes memory data) external returns (uint256 amount0, uint256 amount1);
    function observations(uint256) external view returns (uint32 blockTimestamp, int56 tickCumulative, uint160 secondsPerLiquidityCumulativeX128, bool initialized);
    function observe(uint32[] memory secondsAgos) external view returns (int56[] memory tickCumulatives, uint160[] memory secondsPerLiquidityCumulativeX128s);
    function positions(bytes32) external view returns (uint128 liquidity, uint256 feeGrowthInside0LastX128, uint256 feeGrowthInside1LastX128, uint128 tokensOwed0, uint128 tokensOwed1);
    function protocolFees() external view returns (uint128 token0, uint128 token1);
    function setFeeProtocol(uint32 feeProtocol0, uint32 feeProtocol1) external;
    function setLmPool(address _lmPool) external;
    function slot0() external view returns (uint160 sqrtPriceX96, int24 tick, uint16 observationIndex, uint16 observationCardinality, uint16 observationCardinalityNext, uint32 feeProtocol, bool unlocked);
    function snapshotCumulativesInside(int24 tickLower, int24 tickUpper) external view returns (int56 tickCumulativeInside, uint160 secondsPerLiquidityInsideX128, uint32 secondsInside);
    function swap(address recipient, bool zeroForOne, int256 amountSpecified, uint160 sqrtPriceLimitX96, bytes memory data) external returns (int256 amount0, int256 amount1);
    function tickBitmap(int16) external view returns (uint256);
    function tickSpacing() external view returns (int24);
    function ticks(int24) external view returns (uint128 liquidityGross, int128 liquidityNet, uint256 feeGrowthOutside0X128, uint256 feeGrowthOutside1X128, int56 tickCumulativeOutside, uint160 secondsPerLiquidityOutsideX128, uint32 secondsOutside, bool initialized);
    function token0() external view returns (address);
    function token1() external view returns (address);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "burn",
    "inputs": [
      {
        "name": "tickLower",
        "type": "int24",
        "internalType": "int24"
      },
      {
        "name": "tickUpper",
        "type": "int24",
        "internalType": "int24"
      },
      {
        "name": "amount",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "outputs": [
      {
        "name": "amount0",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "amount1",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "collect",
    "inputs": [
      {
        "name": "recipient",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tickLower",
        "type": "int24",
        "internalType": "int24"
      },
      {
        "name": "tickUpper",
        "type": "int24",
        "internalType": "int24"
      },
      {
        "name": "amount0Requested",
        "type": "uint128",
        "internalType": "uint128"
      },
      {
        "name": "amount1Requested",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "outputs": [
      {
        "name": "amount0",
        "type": "uint128",
        "internalType": "uint128"
      },
      {
        "name": "amount1",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "collectProtocol",
    "inputs": [
      {
        "name": "recipient",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount0Requested",
        "type": "uint128",
        "internalType": "uint128"
      },
      {
        "name": "amount1Requested",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "outputs": [
      {
        "name": "amount0",
        "type": "uint128",
        "internalType": "uint128"
      },
      {
        "name": "amount1",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "factory",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "fee",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint24",
        "internalType": "uint24"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "feeGrowthGlobal0X128",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "feeGrowthGlobal1X128",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "flash",
    "inputs": [
      {
        "name": "recipient",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount0",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "amount1",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "increaseObservationCardinalityNext",
    "inputs": [
      {
        "name": "observationCardinalityNext",
        "type": "uint16",
        "internalType": "uint16"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "initialize",
    "inputs": [
      {
        "name": "sqrtPriceX96",
        "type": "uint160",
        "internalType": "uint160"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "liquidity",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "lmPool",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract IPancakeV3LmPool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "maxLiquidityPerTick",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "mint",
    "inputs": [
      {
        "name": "recipient",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tickLower",
        "type": "int24",
        "internalType": "int24"
      },
      {
        "name": "tickUpper",
        "type": "int24",
        "internalType": "int24"
      },
      {
        "name": "amount",
        "type": "uint128",
        "internalType": "uint128"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [
      {
        "name": "amount0",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "amount1",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "observations",
    "inputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "blockTimestamp",
        "type": "uint32",
        "internalType": "uint32"
      },
      {
        "name": "tickCumulative",
        "type": "int56",
        "internalType": "int56"
      },
      {
        "name": "secondsPerLiquidityCumulativeX128",
        "type": "uint160",
        "internalType": "uint160"
      },
      {
        "name": "initialized",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "observe",
    "inputs": [
      {
        "name": "secondsAgos",
        "type": "uint32[]",
        "internalType": "uint32[]"
      }
    ],
    "outputs": [
      {
        "name": "tickCumulatives",
        "type": "int56[]",
        "internalType": "int56[]"
      },
      {
        "name": "secondsPerLiquidityCumulativeX128s",
        "type": "uint160[]",
        "internalType": "uint160[]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "positions",
    "inputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "liquidity",
        "type": "uint128",
        "internalType": "uint128"
      },
      {
        "name": "feeGrowthInside0LastX128",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "feeGrowthInside1LastX128",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tokensOwed0",
        "type": "uint128",
        "internalType": "uint128"
      },
      {
        "name": "tokensOwed1",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "protocolFees",
    "inputs": [],
    "outputs": [
      {
        "name": "token0",
        "type": "uint128",
        "internalType": "uint128"
      },
      {
        "name": "token1",
        "type": "uint128",
        "internalType": "uint128"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "setFeeProtocol",
    "inputs": [
      {
        "name": "feeProtocol0",
        "type": "uint32",
        "internalType": "uint32"
      },
      {
        "name": "feeProtocol1",
        "type": "uint32",
        "internalType": "uint32"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setLmPool",
    "inputs": [
      {
        "name": "_lmPool",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "slot0",
    "inputs": [],
    "outputs": [
      {
        "name": "sqrtPriceX96",
        "type": "uint160",
        "internalType": "uint160"
      },
      {
        "name": "tick",
        "type": "int24",
        "internalType": "int24"
      },
      {
        "name": "observationIndex",
        "type": "uint16",
        "internalType": "uint16"
      },
      {
        "name": "observationCardinality",
        "type": "uint16",
        "internalType": "uint16"
      },
      {
        "name": "observationCardinalityNext",
        "type": "uint16",
        "internalType": "uint16"
      },
      {
        "name": "feeProtocol",
        "type": "uint32",
        "internalType": "uint32"
      },
      {
        "name": "unlocked",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "snapshotCumulativesInside",
    "inputs": [
      {
        "name": "tickLower",
        "type": "int24",
        "internalType": "int24"
      },
      {
        "name": "tickUpper",
        "type": "int24",
        "internalType": "int24"
      }
    ],
    "outputs": [
      {
        "name": "tickCumulativeInside",
        "type": "int56",
        "internalType": "int56"
      },
      {
        "name": "secondsPerLiquidityInsideX128",
        "type": "uint160",
        "internalType": "uint160"
      },
      {
        "name": "secondsInside",
        "type": "uint32",
        "internalType": "uint32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "swap",
    "inputs": [
      {
        "name": "recipient",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "zeroForOne",
        "type": "bool",
        "internalType": "bool"
      },
      {
        "name": "amountSpecified",
        "type": "int256",
        "internalType": "int256"
      },
      {
        "name": "sqrtPriceLimitX96",
        "type": "uint160",
        "internalType": "uint160"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [
      {
        "name": "amount0",
        "type": "int256",
        "internalType": "int256"
      },
      {
        "name": "amount1",
        "type": "int256",
        "internalType": "int256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "tickBitmap",
    "inputs": [
      {
        "name": "",
        "type": "int16",
        "internalType": "int16"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "tickSpacing",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "int24",
        "internalType": "int24"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "ticks",
    "inputs": [
      {
        "name": "",
        "type": "int24",
        "internalType": "int24"
      }
    ],
    "outputs": [
      {
        "name": "liquidityGross",
        "type": "uint128",
        "internalType": "uint128"
      },
      {
        "name": "liquidityNet",
        "type": "int128",
        "internalType": "int128"
      },
      {
        "name": "feeGrowthOutside0X128",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "feeGrowthOutside1X128",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "tickCumulativeOutside",
        "type": "int56",
        "internalType": "int56"
      },
      {
        "name": "secondsPerLiquidityOutsideX128",
        "type": "uint160",
        "internalType": "uint160"
      },
      {
        "name": "secondsOutside",
        "type": "uint32",
        "internalType": "uint32"
      },
      {
        "name": "initialized",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "token0",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "token1",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "event",
    "name": "Burn",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "tickLower",
        "type": "int24",
        "indexed": true,
        "internalType": "int24"
      },
      {
        "name": "tickUpper",
        "type": "int24",
        "indexed": true,
        "internalType": "int24"
      },
      {
        "name": "amount",
        "type": "uint128",
        "indexed": false,
        "internalType": "uint128"
      },
      {
        "name": "amount0",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "amount1",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Collect",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "recipient",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "tickLower",
        "type": "int24",
        "indexed": true,
        "internalType": "int24"
      },
      {
        "name": "tickUpper",
        "type": "int24",
        "indexed": true,
        "internalType": "int24"
      },
      {
        "name": "amount0",
        "type": "uint128",
        "indexed": false,
        "internalType": "uint128"
      },
      {
        "name": "amount1",
        "type": "uint128",
        "indexed": false,
        "internalType": "uint128"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "CollectProtocol",
    "inputs": [
      {
        "name": "sender",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "recipient",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount0",
        "type": "uint128",
        "indexed": false,
        "internalType": "uint128"
      },
      {
        "name": "amount1",
        "type": "uint128",
        "indexed": false,
        "internalType": "uint128"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Flash",
    "inputs": [
      {
        "name": "sender",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "recipient",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount0",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "amount1",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "paid0",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "paid1",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "IncreaseObservationCardinalityNext",
    "inputs": [
      {
        "name": "observationCardinalityNextOld",
        "type": "uint16",
        "indexed": false,
        "internalType": "uint16"
      },
      {
        "name": "observationCardinalityNextNew",
        "type": "uint16",
        "indexed": false,
        "internalType": "uint16"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Initialize",
    "inputs": [
      {
        "name": "sqrtPriceX96",
        "type": "uint160",
        "indexed": false,
        "internalType": "uint160"
      },
      {
        "name": "tick",
        "type": "int24",
        "indexed": false,
        "internalType": "int24"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Mint",
    "inputs": [
      {
        "name": "sender",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "owner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "tickLower",
        "type": "int24",
        "indexed": true,
        "internalType": "int24"
      },
      {
        "name": "tickUpper",
        "type": "int24",
        "indexed": true,
        "internalType": "int24"
      },
      {
        "name": "amount",
        "type": "uint128",
        "indexed": false,
        "internalType": "uint128"
      },
      {
        "name": "amount0",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "amount1",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "SetFeeProtocol",
    "inputs": [
      {
        "name": "feeProtocol0Old",
        "type": "uint32",
        "indexed": false,
        "internalType": "uint32"
      },
      {
        "name": "feeProtocol1Old",
        "type": "uint32",
        "indexed": false,
        "internalType": "uint32"
      },
      {
        "name": "feeProtocol0New",
        "type": "uint32",
        "indexed": false,
        "internalType": "uint32"
      },
      {
        "name": "feeProtocol1New",
        "type": "uint32",
        "indexed": false,
        "internalType": "uint32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "SetLmPoolEvent",
    "inputs": [
      {
        "name": "addr",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Swap",
    "inputs": [
      {
        "name": "sender",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "recipient",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount0",
        "type": "int256",
        "indexed": false,
        "internalType": "int256"
      },
      {
        "name": "amount1",
        "type": "int256",
        "indexed": false,
        "internalType": "int256"
      },
      {
        "name": "sqrtPriceX96",
        "type": "uint160",
        "indexed": false,
        "internalType": "uint160"
      },
      {
        "name": "liquidity",
        "type": "uint128",
        "indexed": false,
        "internalType": "uint128"
      },
      {
        "name": "tick",
        "type": "int24",
        "indexed": false,
        "internalType": "int24"
      },
      {
        "name": "protocolFeesToken0",
        "type": "uint128",
        "indexed": false,
        "internalType": "uint128"
      },
      {
        "name": "protocolFeesToken1",
        "type": "uint128",
        "indexed": false,
        "internalType": "uint128"
      }
    ],
    "anonymous": false
  }
]

Structs

Burn

Event with signature Burn(address,int24,int24,uint128,uint256,uint256) and selector 0x0c396cd989a39f4459b5fa1aed6a9a8dcdbc45908acfd67e028cd568da98982c.

Collect

Event with signature Collect(address,address,int24,int24,uint128,uint128) and selector 0x70935338e69775456a85ddef226c395fb668b63fa0115f5f20610b388e6ca9c0.

CollectProtocol

Event with signature CollectProtocol(address,address,uint128,uint128) and selector 0x596b573906218d3411850b26a6b437d6c4522fdb43d2d2386263f86d50b8b151.

Flash

Event with signature Flash(address,address,uint256,uint256,uint256,uint256) and selector 0xbdbdb71d7860376ba52b25a5028beea23581364a40522f6bcfb86bb1f2dca633.

IncreaseObservationCardinalityNext

Event with signature IncreaseObservationCardinalityNext(uint16,uint16) and selector 0xac49e518f90a358f652e4400164f05a5d8f7e35e7747279bc3a93dbf584e125a.

Initialize

Event with signature Initialize(uint160,int24) and selector 0x98636036cb66a9c19a37435efc1e90142190214e8abeb821bdba3f2990dd4c95.

Mint

Event with signature Mint(address,address,int24,int24,uint128,uint256,uint256) and selector 0x7a53080ba414158be7ec69b987b5fb7d07dee101fe85488f0853ae16239d0bde.

SetFeeProtocol

Event with signature SetFeeProtocol(uint32,uint32,uint32,uint32) and selector 0xb3159fed3ddfba67bae294599eafe2d0ec98c08bb38e0e5fb87d33154b6e05aa.

SetLmPoolEvent

Event with signature SetLmPoolEvent(address) and selector 0x29983690a85a11696ce8a357993744f8d5a74fde14653e517cc2f8608a7235e9.

Swap

Event with signature Swap(address,address,int256,int256,uint160,uint128,int24,uint128,uint128) and selector 0x19b47279256b2a23a1665c810c8d55a1758940ee09377d4f8d26497a3577dc83.

burnCall

Function with signature burn(int24,int24,uint128) and selector 0xa34123a7.

burnReturn

Container type for the return parameters of the burn(int24,int24,uint128) function.

collectCall

Function with signature collect(address,int24,int24,uint128,uint128) and selector 0x4f1eb3d8.

collectProtocolCall

Function with signature collectProtocol(address,uint128,uint128) and selector 0x85b66729.

collectProtocolReturn

Container type for the return parameters of the collectProtocol(address,uint128,uint128) function.

collectReturn

Container type for the return parameters of the collect(address,int24,int24,uint128,uint128) function.

constructorCall

Constructor`.

factoryCall

Function with signature factory() and selector 0xc45a0155.

factoryReturn

Container type for the return parameters of the factory() function.

feeCall

Function with signature fee() and selector 0xddca3f43.

feeGrowthGlobal0X128Call

Function with signature feeGrowthGlobal0X128() and selector 0xf3058399.

feeGrowthGlobal0X128Return

Container type for the return parameters of the feeGrowthGlobal0X128() function.

feeGrowthGlobal1X128Call

Function with signature feeGrowthGlobal1X128() and selector 0x46141319.

feeGrowthGlobal1X128Return

Container type for the return parameters of the feeGrowthGlobal1X128() function.

feeReturn

Container type for the return parameters of the fee() function.

flashCall

Function with signature flash(address,uint256,uint256,bytes) and selector 0x490e6cbc.

flashReturn

Container type for the return parameters of the flash(address,uint256,uint256,bytes) function.

increaseObservationCardinalityNextCall

Function with signature increaseObservationCardinalityNext(uint16) and selector 0x32148f67.

increaseObservationCardinalityNextReturn

Container type for the return parameters of the increaseObservationCardinalityNext(uint16) function.

initializeCall

Function with signature initialize(uint160) and selector 0xf637731d.

initializeReturn

Container type for the return parameters of the initialize(uint160) function.

liquidityCall

Function with signature liquidity() and selector 0x1a686502.

liquidityReturn

Container type for the return parameters of the liquidity() function.

lmPoolCall

Function with signature lmPool() and selector 0x540d4918.

lmPoolReturn

Container type for the return parameters of the lmPool() function.

maxLiquidityPerTickCall

Function with signature maxLiquidityPerTick() and selector 0x70cf754a.

maxLiquidityPerTickReturn

Container type for the return parameters of the maxLiquidityPerTick() function.

mintCall

Function with signature mint(address,int24,int24,uint128,bytes) and selector 0x3c8a7d8d.

mintReturn

Container type for the return parameters of the mint(address,int24,int24,uint128,bytes) function.

observationsCall

Function with signature observations(uint256) and selector 0x252c09d7.

observationsReturn

Container type for the return parameters of the observations(uint256) function.

observeCall

Function with signature observe(uint32[]) and selector 0x883bdbfd.

observeReturn

Container type for the return parameters of the observe(uint32[]) function.

positionsCall

Function with signature positions(bytes32) and selector 0x514ea4bf.

positionsReturn

Container type for the return parameters of the positions(bytes32) function.

protocolFeesCall

Function with signature protocolFees() and selector 0x1ad8b03b.

protocolFeesReturn

Container type for the return parameters of the protocolFees() function.

setFeeProtocolCall

Function with signature setFeeProtocol(uint32,uint32) and selector 0xb0d0d211.

setFeeProtocolReturn

Container type for the return parameters of the setFeeProtocol(uint32,uint32) function.

setLmPoolCall

Function with signature setLmPool(address) and selector 0xcc7e7fa2.

setLmPoolReturn

Container type for the return parameters of the setLmPool(address) function.

slot0Call

Function with signature slot0() and selector 0x3850c7bd.

slot0Return

Container type for the return parameters of the slot0() function.

snapshotCumulativesInsideCall

Function with signature snapshotCumulativesInside(int24,int24) and selector 0xa38807f2.

snapshotCumulativesInsideReturn

Container type for the return parameters of the snapshotCumulativesInside(int24,int24) function.

swapCall

Function with signature swap(address,bool,int256,uint160,bytes) and selector 0x128acb08.

swapReturn

Container type for the return parameters of the swap(address,bool,int256,uint160,bytes) function.

tickBitmapCall

Function with signature tickBitmap(int16) and selector 0x5339c296.

tickBitmapReturn

Container type for the return parameters of the tickBitmap(int16) function.

tickSpacingCall

Function with signature tickSpacing() and selector 0xd0c93a7c.

tickSpacingReturn

Container type for the return parameters of the tickSpacing() function.

ticksCall

Function with signature ticks(int24) and selector 0xf30dba93.

ticksReturn

Container type for the return parameters of the ticks(int24) function.

token0Call

Function with signature token0() and selector 0x0dfe1681.

token0Return

Container type for the return parameters of the token0() function.

token1Call

Function with signature token1() and selector 0xd21220a7.

token1Return

Container type for the return parameters of the token1() function.

Enums

PancakeSwapV3Calls

Container for all the PancakeSwapV3 function calls.

PancakeSwapV3Events

Container for all the PancakeSwapV3 events.