Enum brontes_classifier::SushiSwapV3::SushiSwapV3Calls

pub enum SushiSwapV3Calls {

    burn(burnCall),
    collect(collectCall),
    collectProtocol(collectProtocolCall),
    factory(factoryCall),
    fee(feeCall),
    feeGrowthGlobal0X128(feeGrowthGlobal0X128Call),
    feeGrowthGlobal1X128(feeGrowthGlobal1X128Call),
    flash(flashCall),
    increaseObservationCardinalityNext(increaseObservationCardinalityNextCall),
    initialize(initializeCall),
    liquidity(liquidityCall),
    maxLiquidityPerTick(maxLiquidityPerTickCall),
    mint(mintCall),
    observations(observationsCall),
    observe(observeCall),
    positions(positionsCall),
    protocolFees(protocolFeesCall),
    setFeeProtocol(setFeeProtocolCall),
    slot0(slot0Call),
    snapshotCumulativesInside(snapshotCumulativesInsideCall),
    swap(swapCall),
    tickBitmap(tickBitmapCall),
    tickSpacing(tickSpacingCall),
    ticks(ticksCall),
    token0(token0Call),
    token1(token1Call),

}

Container for all the SushiSwapV3 function calls.

Variants

burn(burnCall)

collect(collectCall)

collectProtocol(collectProtocolCall)

factory(factoryCall)

fee(feeCall)

feeGrowthGlobal0X128(feeGrowthGlobal0X128Call)

feeGrowthGlobal1X128(feeGrowthGlobal1X128Call)

flash(flashCall)

increaseObservationCardinalityNext(increaseObservationCardinalityNextCall)

initialize(initializeCall)

liquidity(liquidityCall)

maxLiquidityPerTick(maxLiquidityPerTickCall)

mint(mintCall)

observations(observationsCall)

observe(observeCall)

positions(positionsCall)

protocolFees(protocolFeesCall)

setFeeProtocol(setFeeProtocolCall)

slot0(slot0Call)

snapshotCumulativesInside(snapshotCumulativesInsideCall)

swap(swapCall)

tickBitmap(tickBitmapCall)

tickSpacing(tickSpacingCall)

ticks(ticksCall)

token0(token0Call)

token1(token1Call)

Implementations

impl SushiSwapV3Calls

pub const SELECTORS: &'static [[u8; 4]] = _

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants, as they are sorted instead of ordered by definition.