Module brontes_classifier::UniswapV2

Generated by the following Solidity interface…

interface UniswapV2 {
    event Approval(address indexed owner, address indexed spender, uint256 value);
    event Burn(address indexed sender, uint256 amount0, uint256 amount1, address indexed to);
    event Mint(address indexed sender, uint256 amount0, uint256 amount1);
    event Swap(address indexed sender, uint256 amount0In, uint256 amount1In, uint256 amount0Out, uint256 amount1Out, address indexed to);
    event Sync(uint112 reserve0, uint112 reserve1);
    event Transfer(address indexed from, address indexed to, uint256 value);

    function DOMAIN_SEPARATOR() external view returns (bytes32);
    function MINIMUM_LIQUIDITY() external view returns (uint256);
    function PERMIT_TYPEHASH() external view returns (bytes32);
    function allowance(address, address) external view returns (uint256);
    function approve(address spender, uint256 value) external returns (bool);
    function balanceOf(address) external view returns (uint256);
    function burn(address to) external returns (uint256 amount0, uint256 amount1);
    function decimals() external view returns (uint8);
    function factory() external view returns (address);
    function getReserves() external view returns (uint112 _reserve0, uint112 _reserve1, uint32 _blockTimestampLast);
    function initialize(address _token0, address _token1) external;
    function kLast() external view returns (uint256);
    function mint(address to) external returns (uint256 liquidity);
    function name() external view returns (string memory);
    function nonces(address) external view returns (uint256);
    function permit(address owner, address spender, uint256 value, uint256 deadline, uint8 v, bytes32 r, bytes32 s) external;
    function price0CumulativeLast() external view returns (uint256);
    function price1CumulativeLast() external view returns (uint256);
    function skim(address to) external;
    function swap(uint256 amount0Out, uint256 amount1Out, address to, bytes memory data) external;
    function symbol() external view returns (string memory);
    function sync() external;
    function token0() external view returns (address);
    function token1() external view returns (address);
    function totalSupply() external view returns (uint256);
    function transfer(address to, uint256 value) external returns (bool);
    function transferFrom(address from, address to, uint256 value) external returns (bool);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "DOMAIN_SEPARATOR",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MINIMUM_LIQUIDITY",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "PERMIT_TYPEHASH",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "allowance",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "approve",
    "inputs": [
      {
        "name": "spender",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "value",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "balanceOf",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "burn",
    "inputs": [
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "amount0",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "amount1",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "decimals",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint8",
        "internalType": "uint8"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "factory",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getReserves",
    "inputs": [],
    "outputs": [
      {
        "name": "_reserve0",
        "type": "uint112",
        "internalType": "uint112"
      },
      {
        "name": "_reserve1",
        "type": "uint112",
        "internalType": "uint112"
      },
      {
        "name": "_blockTimestampLast",
        "type": "uint32",
        "internalType": "uint32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "initialize",
    "inputs": [
      {
        "name": "_token0",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "_token1",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "kLast",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "mint",
    "inputs": [
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "liquidity",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "name",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "nonces",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "permit",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "spender",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "value",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "deadline",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "v",
        "type": "uint8",
        "internalType": "uint8"
      },
      {
        "name": "r",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "s",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "price0CumulativeLast",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "price1CumulativeLast",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "skim",
    "inputs": [
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "swap",
    "inputs": [
      {
        "name": "amount0Out",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "amount1Out",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "symbol",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "sync",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "token0",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "token1",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "totalSupply",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "transfer",
    "inputs": [
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "value",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferFrom",
    "inputs": [
      {
        "name": "from",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "value",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "Approval",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "spender",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "value",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Burn",
    "inputs": [
      {
        "name": "sender",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount0",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "amount1",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "to",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Mint",
    "inputs": [
      {
        "name": "sender",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount0",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "amount1",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Swap",
    "inputs": [
      {
        "name": "sender",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount0In",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "amount1In",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "amount0Out",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "amount1Out",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "to",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Sync",
    "inputs": [
      {
        "name": "reserve0",
        "type": "uint112",
        "indexed": false,
        "internalType": "uint112"
      },
      {
        "name": "reserve1",
        "type": "uint112",
        "indexed": false,
        "internalType": "uint112"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Transfer",
    "inputs": [
      {
        "name": "from",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "value",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  }
]

Structs

Approval

Event with signature Approval(address,address,uint256) and selector 0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925.

Burn

Event with signature Burn(address,uint256,uint256,address) and selector 0xdccd412f0b1252819cb1fd330b93224ca42612892bb3f4f789976e6d81936496.

DOMAIN_SEPARATORCall

Function with signature DOMAIN_SEPARATOR() and selector 0x3644e515.

DOMAIN_SEPARATORReturn

Container type for the return parameters of the DOMAIN_SEPARATOR() function.

MINIMUM_LIQUIDITYCall

Function with signature MINIMUM_LIQUIDITY() and selector 0xba9a7a56.

MINIMUM_LIQUIDITYReturn

Container type for the return parameters of the MINIMUM_LIQUIDITY() function.

Mint

Event with signature Mint(address,uint256,uint256) and selector 0x4c209b5fc8ad50758f13e2e1088ba56a560dff690a1c6fef26394f4c03821c4f.

PERMIT_TYPEHASHCall

Function with signature PERMIT_TYPEHASH() and selector 0x30adf81f.

PERMIT_TYPEHASHReturn

Container type for the return parameters of the PERMIT_TYPEHASH() function.

Swap

Event with signature Swap(address,uint256,uint256,uint256,uint256,address) and selector 0xd78ad95fa46c994b6551d0da85fc275fe613ce37657fb8d5e3d130840159d822.

Sync

Event with signature Sync(uint112,uint112) and selector 0x1c411e9a96e071241c2f21f7726b17ae89e3cab4c78be50e062b03a9fffbbad1.

Transfer

Event with signature Transfer(address,address,uint256) and selector 0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef.

allowanceCall

Function with signature allowance(address,address) and selector 0xdd62ed3e.

allowanceReturn

Container type for the return parameters of the allowance(address,address) function.

approveCall

Function with signature approve(address,uint256) and selector 0x095ea7b3.

approveReturn

Container type for the return parameters of the approve(address,uint256) function.

balanceOfCall

Function with signature balanceOf(address) and selector 0x70a08231.

balanceOfReturn

Container type for the return parameters of the balanceOf(address) function.

burnCall

Function with signature burn(address) and selector 0x89afcb44.

burnReturn

Container type for the return parameters of the burn(address) function.

decimalsCall

Function with signature decimals() and selector 0x313ce567.

decimalsReturn

Container type for the return parameters of the decimals() function.

factoryCall

Function with signature factory() and selector 0xc45a0155.

factoryReturn

Container type for the return parameters of the factory() function.

getReservesCall

Function with signature getReserves() and selector 0x0902f1ac.

getReservesReturn

Container type for the return parameters of the getReserves() function.

initializeCall

Function with signature initialize(address,address) and selector 0x485cc955.

initializeReturn

Container type for the return parameters of the initialize(address,address) function.

kLastCall

Function with signature kLast() and selector 0x7464fc3d.

kLastReturn

Container type for the return parameters of the kLast() function.

mintCall

Function with signature mint(address) and selector 0x6a627842.

mintReturn

Container type for the return parameters of the mint(address) function.

nameCall

Function with signature name() and selector 0x06fdde03.

nameReturn

Container type for the return parameters of the name() function.

noncesCall

Function with signature nonces(address) and selector 0x7ecebe00.

noncesReturn

Container type for the return parameters of the nonces(address) function.

permitCall

Function with signature permit(address,address,uint256,uint256,uint8,bytes32,bytes32) and selector 0xd505accf.

permitReturn

Container type for the return parameters of the permit(address,address,uint256,uint256,uint8,bytes32,bytes32) function.

price0CumulativeLastCall

Function with signature price0CumulativeLast() and selector 0x5909c0d5.

price0CumulativeLastReturn

Container type for the return parameters of the price0CumulativeLast() function.

price1CumulativeLastCall

Function with signature price1CumulativeLast() and selector 0x5a3d5493.

price1CumulativeLastReturn

Container type for the return parameters of the price1CumulativeLast() function.

skimCall

Function with signature skim(address) and selector 0xbc25cf77.

skimReturn

Container type for the return parameters of the skim(address) function.

swapCall

Function with signature swap(uint256,uint256,address,bytes) and selector 0x022c0d9f.

swapReturn

Container type for the return parameters of the swap(uint256,uint256,address,bytes) function.

symbolCall

Function with signature symbol() and selector 0x95d89b41.

symbolReturn

Container type for the return parameters of the symbol() function.

syncCall

Function with signature sync() and selector 0xfff6cae9.

syncReturn

Container type for the return parameters of the sync() function.

token0Call

Function with signature token0() and selector 0x0dfe1681.

token0Return

Container type for the return parameters of the token0() function.

token1Call

Function with signature token1() and selector 0xd21220a7.

token1Return

Container type for the return parameters of the token1() function.

totalSupplyCall

Function with signature totalSupply() and selector 0x18160ddd.

totalSupplyReturn

Container type for the return parameters of the totalSupply() function.

transferCall

Function with signature transfer(address,uint256) and selector 0xa9059cbb.

transferFromCall

Function with signature transferFrom(address,address,uint256) and selector 0x23b872dd.

transferFromReturn

Container type for the return parameters of the transferFrom(address,address,uint256) function.

transferReturn

Container type for the return parameters of the transfer(address,uint256) function.

Enums

UniswapV2Calls

Container for all the UniswapV2 function calls.

UniswapV2Events

Container for all the UniswapV2 events.