Module brontes_classifier::UniswapV2Factory

Generated by the following Solidity interface…

interface UniswapV2Factory {
    event PairCreated(address indexed token0, address indexed token1, address pair, uint256);

    function allPairs(uint256) external view returns (address);
    function allPairsLength() external view returns (uint256);
    function createPair(address tokenA, address tokenB) external returns (address pair);
    function feeTo() external view returns (address);
    function feeToSetter() external view returns (address);
    function getPair(address, address) external view returns (address);
    function setFeeTo(address _feeTo) external;
    function setFeeToSetter(address _feeToSetter) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_feeToSetter",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "allPairs",
    "inputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "allPairsLength",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "createPair",
    "inputs": [
      {
        "name": "tokenA",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenB",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "pair",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "feeTo",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "feeToSetter",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getPair",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "setFeeTo",
    "inputs": [
      {
        "name": "_feeTo",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setFeeToSetter",
    "inputs": [
      {
        "name": "_feeToSetter",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "PairCreated",
    "inputs": [
      {
        "name": "token0",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "token1",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "pair",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  }
]

Structs

PairCreated

Event with signature PairCreated(address,address,address,uint256) and selector 0x0d3648bd0f6ba80134a33ba9275ac585d9d315f0ad8355cddefde31afa28d0e9.

allPairsCall

Function with signature allPairs(uint256) and selector 0x1e3dd18b.

allPairsLengthCall

Function with signature allPairsLength() and selector 0x574f2ba3.

allPairsLengthReturn

Container type for the return parameters of the allPairsLength() function.

allPairsReturn

Container type for the return parameters of the allPairs(uint256) function.

createPairCall

Function with signature createPair(address,address) and selector 0xc9c65396.

createPairReturn

Container type for the return parameters of the createPair(address,address) function.

feeToCall

Function with signature feeTo() and selector 0x017e7e58.

feeToReturn

Container type for the return parameters of the feeTo() function.

feeToSetterCall

Function with signature feeToSetter() and selector 0x094b7415.

feeToSetterReturn

Container type for the return parameters of the feeToSetter() function.

getPairCall

Function with signature getPair(address,address) and selector 0xe6a43905.

getPairReturn

Container type for the return parameters of the getPair(address,address) function.

setFeeToCall

Function with signature setFeeTo(address) and selector 0xf46901ed.

setFeeToReturn

Container type for the return parameters of the setFeeTo(address) function.

setFeeToSetterCall

Function with signature setFeeToSetter(address) and selector 0xa2e74af6.

setFeeToSetterReturn

Container type for the return parameters of the setFeeToSetter(address) function.

Enums

UniswapV2FactoryCalls

Container for all the UniswapV2Factory function calls.

UniswapV2FactoryEvents

Container for all the UniswapV2Factory events.