Module brontes_classifier::UniswapX

Generated by the following Solidity interface…

interface UniswapX {
    struct SignedOrder {
        bytes order;
        bytes sig;
    }

    error DeadlineBeforeEndTime();
    error DeadlinePassed();
    error DuplicateFeeOutput(address duplicateToken);
    error EndTimeBeforeStartTime();
    error FeeTooLarge(address token, uint256 amount, address recipient);
    error IncorrectAmounts();
    error InputAndOutputDecay();
    error InsufficientEth();
    error InvalidFeeToken(address feeToken);
    error InvalidReactor();
    error NativeTransferFailed();
    error NoExclusiveOverride();
    error OrderEndTimeBeforeStartTime();

    event Fill(bytes32 indexed orderHash, address indexed filler, address indexed swapper, uint256 nonce);
    event OwnershipTransferred(address indexed user, address indexed newOwner);
    event ProtocolFeeControllerSet(address oldFeeController, address newFeeController);

    constructor(address _permit2, address _protocolFeeOwner);

    receive() external payable;

    function execute(SignedOrder memory order) external payable;
    function executeBatch(SignedOrder[] memory orders) external payable;
    function executeBatchWithCallback(SignedOrder[] memory orders, bytes memory callbackData) external payable;
    function executeWithCallback(SignedOrder memory order, bytes memory callbackData) external payable;
    function feeController() external view returns (address);
    function owner() external view returns (address);
    function permit2() external view returns (address);
    function setProtocolFeeController(address _newFeeController) external;
    function transferOwnership(address newOwner) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_permit2",
        "type": "address",
        "internalType": "contract IPermit2"
      },
      {
        "name": "_protocolFeeOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "receive",
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "execute",
    "inputs": [
      {
        "name": "order",
        "type": "tuple",
        "internalType": "struct SignedOrder",
        "components": [
          {
            "name": "order",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "sig",
            "type": "bytes",
            "internalType": "bytes"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "executeBatch",
    "inputs": [
      {
        "name": "orders",
        "type": "tuple[]",
        "internalType": "struct SignedOrder[]",
        "components": [
          {
            "name": "order",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "sig",
            "type": "bytes",
            "internalType": "bytes"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "executeBatchWithCallback",
    "inputs": [
      {
        "name": "orders",
        "type": "tuple[]",
        "internalType": "struct SignedOrder[]",
        "components": [
          {
            "name": "order",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "sig",
            "type": "bytes",
            "internalType": "bytes"
          }
        ]
      },
      {
        "name": "callbackData",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "executeWithCallback",
    "inputs": [
      {
        "name": "order",
        "type": "tuple",
        "internalType": "struct SignedOrder",
        "components": [
          {
            "name": "order",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "sig",
            "type": "bytes",
            "internalType": "bytes"
          }
        ]
      },
      {
        "name": "callbackData",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "feeController",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract IProtocolFeeController"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "permit2",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract IPermit2"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "setProtocolFeeController",
    "inputs": [
      {
        "name": "_newFeeController",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "Fill",
    "inputs": [
      {
        "name": "orderHash",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "filler",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "swapper",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "nonce",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "user",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ProtocolFeeControllerSet",
    "inputs": [
      {
        "name": "oldFeeController",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "newFeeController",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "DeadlineBeforeEndTime",
    "inputs": []
  },
  {
    "type": "error",
    "name": "DeadlinePassed",
    "inputs": []
  },
  {
    "type": "error",
    "name": "DuplicateFeeOutput",
    "inputs": [
      {
        "name": "duplicateToken",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "EndTimeBeforeStartTime",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FeeTooLarge",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "recipient",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "IncorrectAmounts",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InputAndOutputDecay",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InsufficientEth",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidFeeToken",
    "inputs": [
      {
        "name": "feeToken",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "InvalidReactor",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NativeTransferFailed",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NoExclusiveOverride",
    "inputs": []
  },
  {
    "type": "error",
    "name": "OrderEndTimeBeforeStartTime",
    "inputs": []
  }
]

Structs

DeadlineBeforeEndTime

Custom error with signature DeadlineBeforeEndTime() and selector 0x773a6187.

DeadlinePassed

Custom error with signature DeadlinePassed() and selector 0x70f65caa.

DuplicateFeeOutput

Custom error with signature DuplicateFeeOutput(address) and selector 0xfff08303.

EndTimeBeforeStartTime

Custom error with signature EndTimeBeforeStartTime() and selector 0x43133453.

FeeTooLarge

Custom error with signature FeeTooLarge(address,uint256,address) and selector 0x82e75656.

Fill

Event with signature Fill(bytes32,address,address,uint256) and selector 0x78ad7ec0e9f89e74012afa58738b6b661c024cb0fd185ee2f616c0a28924bd66.

IncorrectAmounts

Custom error with signature IncorrectAmounts() and selector 0x7c1f8113.

InputAndOutputDecay

Custom error with signature InputAndOutputDecay() and selector 0xd303758b.

InsufficientEth

Custom error with signature InsufficientEth() and selector 0xa01a9df6.

InvalidFeeToken

Custom error with signature InvalidFeeToken(address) and selector 0xeddf07f5.

InvalidReactor

Custom error with signature InvalidReactor() and selector 0x4ddf4a64.

NativeTransferFailed

Custom error with signature NativeTransferFailed() and selector 0xf4b3b1bc.

NoExclusiveOverride

Custom error with signature NoExclusiveOverride() and selector 0xb9ec1e96.

OrderEndTimeBeforeStartTime

Custom error with signature OrderEndTimeBeforeStartTime() and selector 0x48fee69c.

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0.

ProtocolFeeControllerSet

Event with signature ProtocolFeeControllerSet(address,address) and selector 0xb904ae9529e373e48bc82df4326cceaf1b4c472babf37f5b7dec46fecc6b53e0.

SignedOrder

constructorCall

Constructor`.

executeBatchCall

Function with signature executeBatch((bytes,bytes)[]) and selector 0x0d7a16c3.

executeBatchReturn

Container type for the return parameters of the executeBatch((bytes,bytes)[]) function.

executeBatchWithCallbackCall

Function with signature executeBatchWithCallback((bytes,bytes)[],bytes) and selector 0x13fb72c7.

executeBatchWithCallbackReturn

Container type for the return parameters of the executeBatchWithCallback((bytes,bytes)[],bytes) function.

executeCall

Function with signature execute((bytes,bytes)) and selector 0x3f62192e.

executeReturn

Container type for the return parameters of the execute((bytes,bytes)) function.

executeWithCallbackCall

Function with signature executeWithCallback((bytes,bytes),bytes) and selector 0x0d335884.

executeWithCallbackReturn

Container type for the return parameters of the executeWithCallback((bytes,bytes),bytes) function.

feeControllerCall

Function with signature feeController() and selector 0x6999b377.

feeControllerReturn

Container type for the return parameters of the feeController() function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b.

ownerReturn

Container type for the return parameters of the owner() function.

permit2Call

Function with signature permit2() and selector 0x12261ee7.

permit2Return

Container type for the return parameters of the permit2() function.

setProtocolFeeControllerCall

Function with signature setProtocolFeeController(address) and selector 0x2d771389.

setProtocolFeeControllerReturn

Container type for the return parameters of the setProtocolFeeController(address) function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b.

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

Enums

UniswapXCalls

Container for all the UniswapX function calls.

UniswapXErrors

Container for all the UniswapX custom errors.

UniswapXEvents

Container for all the UniswapX events.